% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{a_amean}
\alias{a_amean}
\title{Weighted arithmetic mean}
\usage{
a_amean(x, w)
}
\arguments{
\item{x}{A numeric vector.}

\item{w}{A vector of numeric weights of the same length as \code{x}.}
}
\value{
The weighted mean as a scalar value
}
\description{
The vector of weights \code{w} is relative since the formula is:
}
\details{
\deqn{ y = \frac{1}{\sum w_i} \sum w_i x_i }

If \code{x} contains \code{NA}s, these \code{x} values and the corresponding \code{w} values are removed before applying the
formula above.
}
\examples{
x <- c(1:10)
w <- c(10:1)
a_amean(x,w)

}
