% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unbalanced_coin.R
\name{Aggregate.unbalanced_coin}
\alias{Aggregate.unbalanced_coin}
\title{Aggregate method for unbalanced coins}
\usage{
\method{Aggregate}{unbalanced_coin}(x, dset, ...)
}
\arguments{
\item{x}{A coin class object.}

\item{dset}{The name of the data set to apply the function to, which should be accessible in \code{.$Data}.}

\item{...}{arguments passed to or from other methods.}
}
\description{
Aggregate method for unbalanced coins
}
