% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{aggregate2.purse}
\alias{aggregate2.purse}
\title{Aggregate indicators}
\usage{
\method{aggregate2}{purse}(
  x,
  dset,
  f_ag = NULL,
  w = NULL,
  f_ag_para = NULL,
  dat_thresh = NULL,
  write_to = NULL
)
}
\arguments{
\item{x}{A purse object}

\item{dset}{The data set to treat in each coin}

\item{f_ag}{The name of an aggregation function, specified as a string}

\item{w}{An optional data frame of weights. If \code{f_ag} does not require or accept weights, set to \code{"none"}.}

\item{f_ag_para}{Optional parameters to pass to \code{f_ag}, other than \code{x} and \code{w}}

\item{dat_thresh}{An optional data availability threshold, specified as a number between 0 and 1. If a row
within an aggregation group has data availability lower than this threshold, the aggregated value for that row will be
\code{NA}. Data availability, for a row \code{x_row} is defined as \code{sum(!is.na(x_row))/length(x_row)}, i.e. the
fraction of non-\code{NA} values.}

\item{write_to}{If specified, writes the aggregated data to \code{.$Data[[write_to]]}. Default \code{write_to = "Aggregated"}.}
}
\value{
An updated purse with new treated data sets added at \code{.$Data$Aggregated} in each coin.
}
\description{
Aggregate indicators
}
\examples{
#
}
