% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ASEM_iData}
\alias{ASEM_iData}
\title{ASEM raw indicator data}
\format{
A data frame with 51 rows and 60 variables.
}
\source{
\url{https://composite-indicators.jrc.ec.europa.eu/asem-sustainable-connectivity/repository}
}
\usage{
ASEM_iData
}
\description{
A data set containing raw values of indicators for 51 countries, groups and denominators. See the ASEM Portal
for further information and detailed description of each indicator. See also \code{vignette("coins")} for the format
of this data.
}
\details{
This data set is in the new v1.0 format.
}
\keyword{datasets}
