% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ASEM_iData_p}
\alias{ASEM_iData_p}
\title{ASEM raw panel data}
\format{
A data frame with 255 rows and 60 variables.
}
\source{
\url{https://composite-indicators.jrc.ec.europa.eu/asem-sustainable-connectivity/repository}
}
\usage{
ASEM_iData_p
}
\description{
This is an artificially-generated set of panel data (multiple observations of indicators over time) that
is included to build the example "purse" class, i.e. to build composite indicators over time. This will
eventually be replaced with a better example, i.e. a real data set.
}
\details{
This data set is in the new v1.0 format.
}
\keyword{datasets}
