% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ASEM_iMeta}
\alias{ASEM_iMeta}
\title{ASEM indicator metadata}
\format{
A data frame with 68 rows and 9 variables
}
\source{
\url{https://bluefoxr.github.io/COINrDoc/coins-the-currency-of-coinr.html#aggregation-metadata}
}
\usage{
ASEM_iMeta
}
\description{
This contains all metadata for ASEM indicators, including names, weights, directions, etc. See the ASEM Portal
for further information and detailed description of each indicator, and
\href{https://bluefoxr.github.io/COINrDoc/coins-the-currency-of-coinr.html#indicator-metadata}{COINr documentation} for the formatting
of this data set.
}
\details{
This data set is in the new v0.8 format.
}
\keyword{datasets}
