% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ASEMAggMeta}
\alias{ASEMAggMeta}
\title{ASEM aggregate metadata}
\format{
A data frame with 8 rows and 9 variables:
}
\source{
\url{https://composite-indicators.jrc.ec.europa.eu/asem-sustainable-connectivity/repository}
}
\usage{
ASEMAggMeta
}
\description{
This contains all the metadata for the aggregate groups, including the names, weights and codes. See the ASEM Portal
for further information and detailed description of each indicator, and
\href{https://bluefoxr.github.io/COINrDoc/coins-the-currency-of-coinr.html#indicator-data}{COINr documentation} for the formatting
of this data set.
}
\keyword{datasets}
