% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ASEMIndData}
\alias{ASEMIndData}
\title{ASEM raw indicator data}
\format{
A data frame with 51 rows and 60 variables.
}
\source{
\url{https://composite-indicators.jrc.ec.europa.eu/asem-sustainable-connectivity/repository}
}
\usage{
ASEMIndData
}
\description{
A data set containing raw values of indicators for 51 countries, groups and denominators. See the ASEM Portal
for further information and detailed description of each indicator, and
\href{https://bluefoxr.github.io/COINrDoc/coins-the-currency-of-coinr.html#indicator-data}{COINr documentation} for the formatting
of this data set.
}
\keyword{datasets}
