% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ASEMIndMeta}
\alias{ASEMIndMeta}
\title{ASEM indicator metadata}
\format{
A data frame with 49 rows and 7 variables:
}
\source{
\url{https://composite-indicators.jrc.ec.europa.eu/asem-sustainable-connectivity/repository}
}
\usage{
ASEMIndMeta
}
\description{
This contains all metadata for ASEM indicators, including names, weights, directions, etc. See the ASEM Portal
for further information and detailed description of each indicator, and COINr documentation for the formatting
of this data set.
}
\keyword{datasets}
