% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coin_treat.R
\name{BoxCox}
\alias{BoxCox}
\title{Box Cox transformation}
\usage{
BoxCox(x, lambda, makepos = TRUE)
}
\arguments{
\item{x}{A vector or column of data to transform}

\item{lambda}{The lambda parameter of the Box Cox transform}

\item{makepos}{If TRUE (default) makes all values positive by subtracting the minimum and adding 1.}
}
\description{
Simple Box Cox, with no optimisation of lambda.
}
