% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_coin_input}
\alias{check_coin_input}
\title{Stop if object is NOT coin class}
\usage{
check_coin_input(x)
}
\arguments{
\item{x}{An object to be checked.}
}
\value{
An error if input is not coin class.
}
\description{
This helper function is used inside other functions.
}
