% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_and_get.R
\name{check_dset.coin}
\alias{check_dset.coin}
\title{Check for named data set in coin}
\usage{
\method{check_dset}{coin}(x, dset, ...)
}
\arguments{
\item{x}{A coin object}

\item{dset}{Name of data set to check}

\item{...}{Not used}
}
\description{
Check for named data set in coin
}
