% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_and_get.R
\name{check_dset.purse}
\alias{check_dset.purse}
\title{Check for named data set in purse}
\usage{
\method{check_dset}{purse}(x, dset, ...)
}
\arguments{
\item{x}{A purse object}

\item{dset}{Name of data set to check}

\item{...}{Not used}
}
\description{
Check for named data set in purse
}
