% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_dset}
\alias{check_dset}
\title{Check for named data set}
\usage{
check_dset(coin, dset)
}
\arguments{
\item{coin}{A coin class object}

\item{dset}{A character string corresponding to a named data set within \code{coin$Data}. E.g. \code{Raw}.}
}
\value{
Error message if \code{dset} not found.
}
\description{
A helper function to check if a named data set is present. If not, will cause an informative error.
}
\examples{
#

}
