% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_fname}
\alias{check_fname}
\title{Check availability of function}
\usage{
check_fname(f_name)
}
\arguments{
\item{f_name}{}
}
\value{
Nothing or error
}
\description{
Checks if a function is available, and returns an error if not.
}
