% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_purse}
\alias{check_purse}
\title{Checks a purse}
\usage{
check_purse(x)
}
\arguments{
\item{x}{An object to check}
}
\value{
Error messages otherwise if successful, nothing.
}
\description{
Makes sure a purse has the expected format
}
\examples{
#

}
