% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coin_datacheck.R
\name{checkData}
\alias{checkData}
\title{Detailed unit data check and screener by data availability}
\usage{
checkData(
  COIN,
  dset = "Raw",
  ind_thresh = 2/3,
  unit_screen = FALSE,
  Force = NULL,
  out2 = "COIN"
)
}
\arguments{
\item{COIN}{The COIN object}

\item{dset}{The data set to be checked/screened}

\item{ind_thresh}{A data availability threshold, which controls both Rule 1 and 2. Default 0.66. Specify as a fraction.}

\item{unit_screen}{Logical: if TRUE, screens any units with indicator data availability < ind_thresh.
This is output in a new data set .$Data$Screened and added as a column in the output table.}

\item{Force}{A data frame with any additional countries to force inclusion or exclusion. First column is ISO code. Second column either "Include" or "Exclude" for each country to force.}

\item{out2}{Where to output the results. If "COIN" (default for COIN input), appends to updated COIN,
otherwise if "df" outputs to data frame.}
}
\value{
An updated COIN object with tables showing missing data, and a filtered list of countries to include in subsequent calculations.
}
\description{
Gives detailed tables of data availability, and optionally screens units based on a data
availability threshold. Units can be optionally "forced" to be included or excluded, making
exceptions for the data availability threshold.
}
\details{
This function currently only supports COINs as inputs, not data frames.
}
\examples{
\dontrun{checkData(COIN)}

}
