% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{codes2names}
\alias{codes2names}
\title{Convert iCodes to iNames}
\usage{
codes2names(coin, iCodes)
}
\arguments{
\item{coin}{A coin}

\item{iCodes}{A vector of iCodes}
}
\value{
Vector of iNames
}
\description{
Convert iCodes to iNames
}
