% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coin_2excel.R
\name{coin_2excel}
\alias{coin_2excel}
\title{Write data to Excel}
\usage{
coin_2excel(COINobj, fname = "COINresults.xlsx")
}
\arguments{
\item{COINobj}{A COIN object}

\item{fname}{The file name to write to}
}
\value{
An Excel workbook with each table on a separate named tab.
}
\description{
Takes the COIN object and writes all main data tables and other things to an Excel file. At the moment this is just a placeholder to be developed further.
}
\examples{
\dontrun{coin_2excel(COINobj, fname="COINresults.xlsx")}

}
