% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coin_aggregate.R
\name{coin_aggregate}
\alias{coin_aggregate}
\title{Aggregate indicators}
\usage{
coin_aggregate(
  COINobj,
  agtype = "arith_mean",
  agweights = NULL,
  dset = "Normalised",
  agtype_bylevel = NULL
)
}
\arguments{
\item{COINobj}{COIN object}

\item{agtype}{The type of aggregation method.}

\item{agweights}{A list of weights to use in the aggregation. If not specified,
will look in the COIN object (i.e. via coin_weight.R).}

\item{dset}{Which data set (contained in COIN object) to use}

\item{agtype_bylevel}{A character vector with aggregation types for each level}
}
\value{
An updated COIN object containing the new aggregated data set.
}
\description{
Takes indicator data and a specified structure and hierarchically aggregates to a single index (or whatever the structure specified)
}
\examples{
\dontrun{COINlist <- coin_aggregate(COINlist, agtype="arith_mean", dset = "normalised")}

}
