% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coin_auxillary.R
\name{coin_aux_objcheck}
\alias{coin_aux_objcheck}
\title{Format check auxiliary function}
\usage{
coin_aux_objcheck(COINobj, dset = "raw", inames = NULL)
}
\arguments{
\item{COINobj}{Can be either the COIN object, or a data frame.}

\item{dset}{If input is a COIN object, this specifies which data set in .$Data to use.}

\item{inames}{An optional character vector of indicator codes to subset the indicator data. Usefully, can also refer to
an aggregation group name, and data will be subsetted accordingly. NOTE does not work with multiple aggregate group names.}
}
\value{
A list with .$ind_names is a vector with indicator names,
.$ind_data is a data frame /tibble with indicator data, plus possibly grouping/denominator columns, etc.
.$ind_data_only is a data frame/tibble with only indicator data columns (no names, groups, etc)
.$otype is a string, either "COINobj" or "df" describing the input object type
}
\description{
Checks whether function input is COIN object or data frame, and returns input useful to most COINr functions.
The purpose is that most COINr functions are able to take either a COIN object or data frame as input.
Also includes subsetting of indicator data using "inames", which also works by referring to aggregate groups.
}
\examples{
\dontrun{out <- coin_aux_objcheck(COINobj, dset = "raw", inames = NULL)}

}
