% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coin_datacheck.R
\name{coin_datacheck}
\alias{coin_datacheck}
\title{Detailed unit data check and screener by data availability}
\usage{
coin_datacheck(
  COINobj,
  dset = "Raw",
  ind_thresh = 2/3,
  unit_screen = FALSE,
  Force = NULL
)
}
\arguments{
\item{COINobj}{The GII object}

\item{dset}{The data set to be checked/screened}

\item{ind_thresh}{A data availability threshold, which controls both Rule 1 and 2. Default 0.66. Specify as a fraction.}

\item{unit_screen}{Logical: if TRUE, screens any units with indicator data availability < ind_thresh.
This is output in a new data set .$Data$Screened and added as a column in the output table.}

\item{Force}{A data frame with any additional countries to force inclusion or exclusion. First column is ISO code. Second column either "Include" or "Exclude" for each country to force.}
}
\value{
An updated COINobj object with tables showing missing data, and a filtered list of countries to include in subsequent calculations.
}
\description{
Screens countries based on data availability, according to rules used by the GII.
Rule 1: BOTH Input and Output indicators must have data availability greater than a threshold, set as
66% by default, but can be changed using the "ind_thresh" parameter. Rule 1 is applied by default.
Rule 2: All pillars must have at least 66% of their sub-pillars that have 66% indicator data availability.
The 66% threshold is also alternatively specified by "ind_thresh". Rule 2 is NOT applied by default.
}
\examples{
\dontrun{coin_datacheck(COINobj)}

}
