% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coin_impute.R
\name{coin_impute}
\alias{coin_impute}
\title{Impute missing data}
\usage{
coin_impute(
  COINobj,
  imtype = "ind_mean",
  inames = NULL,
  dset = "Raw",
  groupvar = NULL,
  byyear = FALSE
)
}
\arguments{
\item{COINobj}{A dataframe of indicator data following the COINR format}

\item{imtype}{The type of normalisation method. Either "minmax", "zscore", or "custom".}

\item{inames}{Character vector of indiator names, indicating which columns to normalise. Use this if you only want to normalise certain columns, or you are inputting a data frame with some columns which are not to be normalised (e.g. country names, groups, ...)}

\item{dset}{The data set in .$Data to impute}

\item{groupvar}{The name of the column to use for by-group imputation. Only applies when imtype is set to a group option.}

\item{byyear}{Logical: set to TRUE to impute separately for each year, otherwise FALSE to impute across all years.}
}
\value{
A dataframe of normalised indicators.
}
\description{
Imputation of data sets using a variety of methods.
}
\examples{
\dontrun{COINobj <- coin_impute(COINobj, imtype = "ind_mean", dset = "raw")}

}
