% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coin_indicatordash.R
\name{coin_indicatordash}
\alias{coin_indicatordash}
\title{Indicator visualisation dashboard}
\usage{
coin_indicatordash(COINobj, inames = NULL, dset = "raw")
}
\arguments{
\item{COINobj}{The COIN object, or a data frame of indicator data.}

\item{inames}{A set of indicator codes to include. Defaults to all indicators.}

\item{dset}{The data set to plot.}
}
\value{
Interactive visualisation
}
\description{
Generates an interactive visualisation of one or two indicators at a time. Requires Shiny and an active R session.
}
\examples{
\dontrun{coin_indicatordash(COINobj, inames = NULL, dset = "raw")}

}
