% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coin_plotcorr.R
\name{coin_plotcorr}
\alias{coin_plotcorr}
\title{Plots of correlations}
\usage{
coin_plotcorr(COINobj, inames = NULL, dset = "Raw", plotgroups = TRUE)
}
\arguments{
\item{COINobj}{A list of indicator data, stuctured using the COIN_assemble function}

\item{inames}{An optional character vector of indicator codes to plot}

\item{dset}{The source dataset to select}

\item{plotgroups}{Logical: if TRUE, overlays rectangles representing aggregation groups}
}
\value{
A correlation plot.
}
\description{
Visualisation of correlations between indicators.
}
\examples{
\dontrun{coin_plotcorr(COINobj)}

}
