% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coin_plotframework.R
\name{coin_plotframework}
\alias{coin_plotframework}
\title{Interactive sunburst plot of index structure}
\usage{
coin_plotframework(COINobj)
}
\arguments{
\item{COINobj}{COIN object, or list with first entry is the indicator metadata, second entry is the aggregation metadata}
}
\value{
Interactive sunburst plot.
}
\description{
Plots the structure of the index using a sunburst plot. Output can be used as an interactive plot
in html documents, e.g. via R Markdown.
}
\examples{
\dontrun{coin_plotframework(COINobj)}

}
