% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coin_rew8r.R
\name{coin_rew8r}
\alias{coin_rew8r}
\title{Reweight indicators}
\usage{
coin_rew8r(COINobj)
}
\arguments{
\item{COINobj}{COIN object}
}
\value{
An updated COIN object with additional sets of weights (if saved from the app).
}
\description{
Interactive gadget which lets you adjust weights of indicators and see the effects on the results and
the correlations. Weights can be saved with new names to the COIN object.
}
