% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coin_treat.R
\name{coin_treat}
\alias{coin_treat}
\title{Treatment of outliers}
\usage{
coin_treat(
  COINobj,
  dset = "Raw",
  winmax = NULL,
  winchange = FALSE,
  deflog = "log",
  t_skew = 2,
  t_kurt = 3.5,
  individual = NULL,
  indiv_only = TRUE
)
}
\arguments{
\item{COINobj}{The COIN object}

\item{dset}{The data set to treat}

\item{winmax}{The maximum number of points to Winsorise for each indicator. If NA, will keep Winsorising until skew&kurt thresholds achieved (but it is likely this will cause errors)}

\item{winchange}{Logical: if TRUE, Winsorisation can change direction from one iteration to the next. Otherwise if FALSE (default), no change.}

\item{deflog}{If "log", use simple ln(x) as log transform. If "GIIlog", use GII log transformation.}

\item{t_skew}{Absolute skew threshold (default 2)}

\item{t_kurt}{Kurtosis threshold (default 3.5)}

\item{individual}{A data frame specifying individual treatment for each indicator}

\item{indiv_only}{Logical: if TRUE, only the indicators specified in "individual" are treated.
If false, all indicators are treated: any outside of "individual" will get default treatment.}
}
\value{
A treated data set plus information about how the data was treated.
}
\description{
Takes the COIN object and Winsorises indicators where necessary or specified, or reverts to log transform.
}
