% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coinToolIn.R
\name{COINToolIn}
\alias{COINToolIn}
\title{Import data directly from COIN Tool}
\usage{
COINToolIn(fname)
}
\arguments{
\item{fname}{The file name and path to read, e.g. "C:/Documents/COINToolFile.xlsx"}
}
\value{
A list
}
\description{
This provides a direct interface for reading a COIN Tool input deck and converting it to COINr.
You need to provide a COIN Tool file, with the "Database" sheet properly compiled.
}
\examples{
\dontrun{
out <- coinToolIn("C:/Documents/COINToolFile.xlsx")}

}
