% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coin_resultsdash.R
\name{colourTable}
\alias{colourTable}
\title{Conditionally formatted table}
\usage{
colourTable(
  df,
  freeze1 = TRUE,
  sortcol = NULL,
  sortorder = "desc",
  searchable = TRUE,
  pagesize = 10,
  cell_colours = NULL,
  reverse_colours = FALSE
)
}
\arguments{
\item{df}{The COIN object, or a data frame of indicator data.}

\item{freeze1}{If TRUE (default), freezes the first column. This may be for example the unit name or code.}

\item{sortcol}{A column name to sort the table by. Defaults to first numerical column. Set to "none" to disable.}

\item{sortorder}{Either "desc" for sorted col to be sorted from high to low (default) or "asc" for the opposite.}

\item{searchable}{If TRUE, includes a search box}

\item{pagesize}{The number of rows to display on each page.}

\item{cell_colours}{A character vector of colour codes (e.g. Hex codes) to use for the colour palette. Should be in
order of low to high values. Defaults to a simple green palette. See grDevices::colorRamp() for more info.}

\item{reverse_colours}{If TRUE, reverses the colour map- useful for rank tables where lowest numbers mean high scores.}
}
\value{
Interactive table
}
\description{
Given a data frame, generates a conditionally-formatted html table using reactable. This function is used by
iplotTable(). It is a quick wrapper for \code{reactable::reactable}.
}
\examples{
\dontrun{colourTable(as.data.frame(matrix(runif(12), 3, 4)))}

}
\seealso{
\itemize{
\item \code{\link[=iplotTable]{iplotTable()}} Interactive table of indicator data (from a COIN)
}
}
