% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coin_compare.R
\name{compareCOINs}
\alias{compareCOINs}
\title{Interactive comparison of COINs}
\usage{
compareCOINs(COINbase, ... = NULL)
}
\arguments{
\item{COINbase}{A named list of COIN versions, or a single COIN version}

\item{...}{Further versions of the COIN.}
}
\value{
App
}
\description{
Compares two different COIN objects in terms of scores and ranks.
}
\examples{
\dontrun{
compareCOINs(COIN1, COIN2, COIN3)
compareCOINs(list(COIN_nominal = COIN1, COIN_alt1 = COIN2))
}

}
