% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coin_resultstable.R
\name{compareDF}
\alias{compareDF}
\title{Compare two data frames}
\usage{
compareDF(df1, df2, matchcol, sigfigs = 5)
}
\arguments{
\item{df1}{A data frame}

\item{df2}{Another data frame}

\item{matchcol}{A common column name that is used to match row order. E.g. this might be UnitCode.}

\item{sigfigs}{The number of significant figures to use for matching numerical columns}
}
\value{
A list with comparison results
}
\description{
A custom function for comparing two data frames of indicator data, to see whether they match up, at a specified number of
significant figures.
}
\details{
This function compares numerical and non-numerical columns to see if they match. Rows and columns can be in any order. The function
performs the following checks:
\itemize{
\item Checks that the two data frames are the same size
\item Checks that column names are the same, and that the matching column has the same entries
\item Checks column by column that the elements are the same, after sorting according to the matching column
}

It then returns summarises for each column whether there are any differences, and also what the differences are, if any.

This is intended to cross-check results. For example, if you run something in COINr and want to check indicator results against
external calculations.
}
\examples{
\dontrun{"test")}

}
