% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coin_compare.R
\name{compTable}
\alias{compTable}
\title{Rank comparison table between 2 COINs}
\usage{
compTable(
  COIN1,
  COIN2,
  dset = "Raw",
  isel,
  COINnames = NULL,
  sort_by = "AbsRankChange"
)
}
\arguments{
\item{COIN1}{First COIN obj}

\item{COIN2}{Second COIN obj}

\item{dset}{The data set of interest}

\item{isel}{The indicator/col of interest}

\item{COINnames}{An optional character vector of the names of COIN1 and COIN2, to be used in the table headers.}

\item{sort_by}{If "RankCOIN1", sorts by the indicator values of COIN1, if "RankCOIN2", sorts by COIN2,
if "RankChange", sorts by rank change, and if "AbsRankChange" sorts by absolute rank change.}
}
\value{
A data frame with ranks and rank changes
}
\description{
Takes two COINs, and generates a rank comparison between specified indicator/aggregates. COINs must share at least some common
unit codes, and the indicator selected by \code{isel}.
}
\examples{
\dontrun{
ASEM <- build_ASEM()
# Make a copy
ASEMAltNorm <- ASEM
# Edit .$Method
ASEMAltNorm$Method$normalise$ntype <- "borda"
# Regenerate
ASEMAltNorm <- COINr::regen(ASEMAltNorm, quietly = TRUE)
# compare
compTable(ASEM, ASEMAltNorm, dset = "Aggregated", isel = "Index")
}

}
