% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coin_compare.R
\name{compTableMulti}
\alias{compTableMulti}
\title{Rank tables between multiple COINs}
\usage{
compTableMulti(COINs, tabtype = "Ranks", ibase = 1)
}
\arguments{
\item{COINs}{A list of COINs}

\item{tabtype}{The type of table to generate}

\item{ibase}{The index of the COIN to use as a base comparison}
}
\value{
Table
}
\description{
Takes multiple COINs, and generates a rank comparison at the index level
}
