% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{cronbach_alpha}
\alias{cronbach_alpha}
\title{Calculate Cronbach's Alpha}
\usage{
cronbach_alpha(x)
}
\arguments{
\item{x}{A data frame of indicator data, optionally with \code{uCode} column.}
}
\value{
Cronbach alpha value
}
\description{
Calculates Cronbach's Alpha, given a data frame with all numeric columns
}
\examples{
# Calculate Cronbach's alpha for mtcars
cronbach_alpha(mtcars[, 1:4])

}
