% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treat.R
\name{CTlog}
\alias{CTlog}
\title{Log-transform a vector}
\usage{
CTlog(x, na.rm = FALSE)
}
\arguments{
\item{x}{A numeric vector.}

\item{na.rm}{Set \code{TRUE} to remove \code{NA} values, otherwise returns \code{NA}.}
}
\value{
A log-transformed vector of data.
}
\description{
Performs a log transform on a numeric vector.
}
\details{
Specifically, this performs a modified "COIN Tool log" transform: \code{log(x-min(x) + a)}, where
\code{a <- 0.01*(max(x)-min(x))}.
}
\examples{
x <- runif(20)
CTlog(x)

}
