% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treat.R
\name{CTlog_orig}
\alias{CTlog_orig}
\title{Log-transform a vector}
\usage{
CTlog_orig(x, na.rm = FALSE)
}
\arguments{
\item{x}{A numeric vector.}

\item{na.rm}{Set \code{TRUE} to remove \code{NA} values, otherwise returns \code{NA}.}
}
\value{
A log-transformed vector of data.
}
\description{
Performs a log transform on a numeric vector.
}
\details{
Specifically, this performs a "COIN Tool log" transform: \code{log(x-min(x) + 1)}.
}
\examples{
x <- runif(20)
CTlog_orig(x)

}
