% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom.R
\name{Custom}
\alias{Custom}
\title{Custom operation}
\usage{
Custom(x, ...)
}
\arguments{
\item{x}{Object to be operated on (coin or purse)}

\item{...}{arguments passed to or from other methods.}
}
\value{
Modified object.
}
\description{
Allows a custom data operation on coins or purses.
}
