% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity.R
\name{edit_coin}
\alias{edit_coin}
\title{Edit objects inside a coin}
\usage{
edit_coin(coin, address, new_value, checks = TRUE)
}
\arguments{
\item{coin}{A coin}

\item{address}{A string specifying the location in the coin of the object to edit. This should begin with \code{"$"}, omitting the coin itself
in the address. E.g. if you target \code{coin$x$y$z} enter \code{"$x$y$z"}.}

\item{new_value}{The new value to assign at \code{address}.}

\item{checks}{Logical: if \code{TRUE}, runs some basic checks, otherwise omitted if \code{FALSE}. Setting \code{FALSE} may speed
things up a bit in sensitivity analysis, for example.}
}
\value{
An updated coin
}
\description{
Changes the object found at \code{address} to \code{new_value}.
}
\examples{
#
}
