% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{export_to_excel}
\alias{export_to_excel}
\title{Export a coin to Excel}
\usage{
export_to_excel(coin, fname = "coin_export.xlsx", include_log = FALSE)
}
\arguments{
\item{coin}{A coin class object}

\item{fname}{The file name/path to write to, as a character string}

\item{include_log}{Logical: if \code{TRUE}, also writes data frames from the \code{.$Log} list inside the coin.}
}
\value{
.xlsx file at specified path
}
\description{
Exports the contents of the coin to Excel. This writes all data frames inside the coin to Excel, with each data
frame on a separate tab. Tabs are named according to the position in the coin object. You can write other
data frames by simply attaching them to the coin object somewhere.
}
\examples{
#

}
