% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_and_get.R
\name{extract_iData}
\alias{extract_iData}
\title{Exract things from iData}
\usage{
extract_iData(coin, iData, GET)
}
\arguments{
\item{coin}{A coin}

\item{iData}{An iData format data frame}

\item{GET}{What to get}
}
\value{
Data frame of indicator data, indexed also by time if input is a purse.
}
\description{
A helper function to separate indicator cols from metadata columns in iMeta.
}
\examples{
#

}
