% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coin_aggregate.R
\name{geoMean}
\alias{geoMean}
\title{Weighted geometric mean}
\usage{
geoMean(x, w = NULL)
}
\arguments{
\item{x}{A numeric vector of positive values.}

\item{w}{A vector of weights, which should have length equal to length(x). Weights are relative
and will be rescaled to sum to 1. If w is not specified, defaults to equal weights.}
}
\value{
Geometric mean
}
\description{
Weighted geometric mean of a vector. NAs are skipped by default.
}
