% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_and_get.R
\name{get_data}
\alias{get_data}
\title{Get subsets of indicator data}
\usage{
get_data(x, ...)
}
\arguments{
\item{x}{A coin or purse}

\item{...}{Arguments passed to methods}
}
\value{
Data frame of indicator data, indexed also by time if input is a purse.
}
\description{
A helper function to retrieve a named data set from coin or purse objects.
}
\examples{
#

}
