% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screen_units.R
\name{get_data_avail.data.frame}
\alias{get_data_avail.data.frame}
\title{Get data availability of units}
\usage{
\method{get_data_avail}{data.frame}(x, ...)
}
\arguments{
\item{x}{A data frame}

\item{...}{arguments passed to or from other methods.}
}
\value{
A data frame of data availability statistics for each column of \code{x}.
}
\description{
Returns a data frame of the data availability of each unit (row), as well as percentage of zeros. This
function ignores any non-numeric columns, and returns a data availability table with non-numeric columns
appended at the beginning.
}
\examples{
# data availability of "airquality" data set
get_data_avail(airquality)

}
