% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screen_units.R
\name{get_datAvail.coin}
\alias{get_datAvail.coin}
\title{Get data availability of units}
\usage{
\method{get_datAvail}{coin}(x, dset, out2 = "coin")
}
\value{

}
\description{
Returns a data frame of the data availability of each unit (row) in a given data set, as well as percentage of zeros. This
function ignores any non-numeric columns, and returns a data availability table of numeric columns with non-numeric columns
appended at the beginning.
}
\examples{
#
}
