% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screen_units.R
\name{get_datAvail.data.frame}
\alias{get_datAvail.data.frame}
\title{Get data availability of units}
\usage{
\method{get_datAvail}{data.frame}(x)
}
\value{

}
\description{
Returns a data frame of the data availability of each unit (row), as well as percentage of zeros. This
function ignores any non-numeric columns, and returns a data availability table with non-numeric columns
appended at the beginning.
}
\examples{
#
}
