% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_dea.R
\name{get_DEA}
\alias{get_DEA}
\alias{get_DEA.coin}
\alias{get_DEA.unbalanced_coin}
\title{Re-aggregate with DEA}
\usage{
get_DEA(coin, level = NULL, wr_type = 0, wr_bounds = NULL, wr_rhs = NULL)

\method{get_DEA}{coin}(coin, level = NULL, wr_type = 0, wr_bounds = NULL, wr_rhs = NULL)

\method{get_DEA}{unbalanced_coin}(coin, level = NULL, wr_type = 0, wr_bounds = NULL, wr_rhs = NULL)
}
\arguments{
\item{coin}{A coin-class object.}

\item{level}{The Level in which we wish to re-aggregate the coin using DEA. Defaults to the final aggregation Level. If specified as a lower level,
then it will aggregate all indicators at that Level in a composite. See Details.}

\item{wr_type}{A single numeric value indicating the type of weight restrictions to include to the model. Can take the values: 0, 1, 2, 3, and 4.
Defaults to 0 (no additional restrictions). See \code{\link[=a_dea]{a_dea()}} for further details.}

\item{wr_bounds}{A numeric vector or a matrix containing information on the weight restrictions. See \code{\link[=a_dea]{a_dea()}} for further details.}

\item{wr_rhs}{A numeric vector containing the values of the right-hand side of different forms of weight restrictions. To be used when \code{wr_type} is set to 1.
Defaults to \code{NULL} which indicates a zero right hand side. See \code{\link[=a_dea]{a_dea()}} for further details.}
}
\value{
A list containing the following:
\itemize{
\item a data frame named \code{DEA_CI} containing the unit names, the Pillars to be aggregated, the nominal aggregated Index, and the DEA re-aggregated Index
in a column named "Dea".
\item a data frame \code{DEA_CI_ranks} containing the ranks of the units based on the DEA re-aggregated Index.
\item a data frame named \code{dea_weights} containing the aggregation weights selected by each unit.
\item a data frame named \code{normalised_weights}containing the average normalised weights. These are obtained by averaging
across rows of the \code{dea_weights} data frame and then dividing each average term by the sum of the average terms.
}
}
\description{
Performs the aggregation in the last aggregation Level (Pillars to Index) using the non-parametric flexible-weighting
Benefit-if-the-Doubt (BoD) model. This is an alternative aggregation technique used in Auditing Composite Indices by
CC-COIN. The function must take as a minimum an input \code{coin} which is a coin-class object developed through COINr package
that has been aggregated and so includes a dataset called "Aggregated". If no other parameters are specified, this will
estimate the conventional form of the BoD model, in which no additional restrictions on the aggregation weights are
placed. Such restrictions can be placed through parameters inherited from \code{\link[=a_dea]{a_dea()}}, see Details below.
}
\examples{

# example #1 (no additional weight restrictions)
# build example coin up to Aggregated dataset.
coin <- build_example_coin(up_to = "Aggregate")
# re-aggregate Index
dea_reagg <- get_DEA(coin)

# example #2 (with additional weight restrictions, type 2 restrictions in the
# contribution of each Pillar to the index, max 75\%, min 5\%)
dea_reagg <- get_DEA(coin, wr_type = 2, wr_bounds = c(0.05, 0.75))

# example #3 (aggregation at a lower level)
# In the ASEM data, Level 3 is the final aggregation Level where the two Pillars
# get aggregated to ASEM. Specifying `level = 2` will aggregate the
# eight indicators (five under Connectivity Pillar and three under Syst Pillar)
# to an overall composite.
dea_reagg <- get_DEA(coin, level = 2, wr_type = 2, wr_bounds = c(0.05, 0.75))


}
