% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_dset}
\alias{get_dset}
\title{Gets a named data set and performs checks}
\usage{
get_dset(coin, dset)
}
\arguments{
\item{coin}{A coin class object}

\item{dset}{A character string corresponding to a named data set within \code{coin$Data}. E.g. \code{Raw}.}
}
\value{
Data frame of indicator data.
}
\description{
A helper function to retrieve a named data set from the coin object. Also performs input checks at the
same time.
}
\examples{
#

}
