% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_iCodes_in_group}
\alias{get_iCodes_in_group}
\title{Find children or parents of an iCode}
\usage{
get_iCodes_in_group(coin, iCode_group, at_level)
}
\arguments{
\item{coin}{A coin}

\item{iCode_group}{An iCode to find children/parents for}

\item{at_level}{Level of framework to identify children/parents at}
}
\value{
Character vector of iCodes at the specified level that are children/parents of \code{iCode_group}
}
\description{
Finds the children or parents of any iCode \code{iCode_group} at a specified level \code{at_level}.
This is useful for identifying which indicators belong to a particular pillar or sub-index,
or for finding the parent groups of a set of indicators.
}
\details{
The function works by using the hierarchical structure stored in \code{coin$Meta$Lineage} to
identify all indicators that share the same parent at the specified levels. It works with
both balanced and unbalanced frameworks.
}
\examples{
# Build example coin
coin <- build_example_coin(up_to = "new_coin", quietly = TRUE)

# Find all level 2 indicators (pillars) within the "Conn" sub-index (level 3)
get_iCodes_in_group(coin, iCode_group = "Conn", at_level = 2)

# Find all level 1 indicators within the "Political" pillar
get_iCodes_in_group(coin, iCode_group = "Political", at_level = 1)
}
