% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_coin.R
\name{get_lineage}
\alias{get_lineage}
\title{Make wide structure table}
\usage{
get_lineage(iMeta, level_names = NULL)
}
\arguments{
\item{iMeta}{A data frame of indicator metadata. For specs see \code{\link[=check_iMeta]{check_iMeta()}}.}
}
\value{
Lineage table as data frame
}
\description{
Takes an iMeta table and outputs a wide format index structure, i.e. a table with one column per
level in the index. This is used in later functions to look up the full "ancestry" of any element
in the index.
}
\examples{
get_lineage(ASEM_iMeta)

}
