% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_and_get.R
\name{get_names}
\alias{get_names}
\title{Get names from codes}
\usage{
get_names(coin, uCodes = NULL, iCodes = NULL)
}
\arguments{
\item{coin}{A coin}

\item{uCodes}{Some uCodes}

\item{iCodes}{Some iCodes}
}
\value{
Vector of names
}
\description{
Given either uCodes or iCodes, returns uNames or iNames
}
\examples{
#

}
