% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity.R
\name{get_noisy_weights2}
\alias{get_noisy_weights2}
\title{Noisy replications of weights}
\usage{
get_noisy_weights2(
  w,
  noise_specs = NULL,
  individual_specs = NULL,
  Nrep,
  correct_uniform_dist = FALSE,
  uniform_tol = 0.01
)
}
\arguments{
\item{w}{A data frame of weights, in the format found in \code{.$Meta$Weights}.}

\item{noise_specs}{a data frame with columns:
\itemize{
\item \code{Level}: The aggregation level to apply noise to
\item \code{NoiseFactor}: The size of the perturbation: setting e.g. 0.2 perturbs by +/- 20\% of nominal values.
}}

\item{individual_specs}{Optional list for specifying specific noise factors on individual indicators. A named
list where names are the iCodes of indicators/aggregates, and the entries are the noise factors. This overrides
the \code{noise_specs} specifications for those indicators (if there were any).}

\item{Nrep}{The number of weight replications to generate.}

\item{correct_uniform_dist}{Logical: if \code{TRUE}, calls \code{get_perturbed_weight_sample()} to generate more-uniform weight
samples via rejection sampling.}

\item{uniform_tol}{Tolerance parameter passed to \code{get_perturbed_weight_sample()}.}
}
\value{
A list of \code{Nrep} sets of weights (data frames).
}
\description{
Given a data frame of weights, this function returns multiple replicates of the weights, with added
noise. This is intended for use in uncertainty and sensitivity analysis. \emph{NOTE:} this function is a modified
version of \code{\link[=get_noisy_weights]{get_noisy_weights()}}.
}
\details{
Weights are expected to be in a data frame format with columns \code{Level}, \code{iCode}, \code{Weight} and \code{Parent} as
used in \code{iMeta}.

Noise is added using two arguments. The first is the \code{noise_specs} argument, which is specified by a data frame with columns
\code{Level} and \code{NoiseFactor}. The aggregation level refers to number of the aggregation level to target
while the \code{NoiseFactor} refers to the size of the perturbation. If e.g. a row is \code{Level = 1} and
\code{NoiseFactor = 0.2}, this will allow the weights in aggregation level 1 to deviate by +/- 20\% of their
nominal values (the values in \code{w}). Note that due to rescaling to sum to 1, this tends to result in
truncated normal distributions for each weight, unless \code{correct_uniform_dist = TRUE} (see below).

If you need more control over the noise applied to individual indicators, use the \code{individual_specs} argument.
Any noise specifications here will override those in \code{noise_specs} if they are specified in both. Keep in
mind that if \code{correct_uniform_dist = TRUE}, it is not possible to perturb only one weight within a group. This
is because it is impossible to vary a single value while keeping the sum near 1.

Finally, the \code{correct_uniform_dist} argument allows to switch to use the \code{get_perturbed_weight_sample()} function
which returns weight samples with more uniform distributions (as opposed to the truncated normal distribution as
mentioned previously).

See the examples vignette for some discussion and demos on the distributions of weights.
}
\examples{
# build example coin
coin <- build_example_coin(up_to = "new_coin", quietly = TRUE)

# get nominal weights
w_nom <- coin$Meta$Ind[coin$Meta$Ind$Type \%in\% c("Indicator", "Aggregate"),
                       c("iCode", "Weight", "Level", "Parent")]

# build data frame specifying the levels to apply the noise at
# here we vary at levels 2 and 3
noise_specs = data.frame(Level = c(2,3),
                         NoiseFactor = c(0.25, 0.25))

# get 100 replications
noisy_wts <- get_noisy_weights2(w = w_nom, noise_specs = noise_specs, Nrep = 100)

# examine one of the noisy weight sets, last few rows
tail(noisy_wts[[1]])

## Example with individual specs for individual components

# specify for two components
individual_specs <- list(Physical = 1, P2P = 0.75)
# run
noisy_wts <- get_noisy_weights2(w = w_nom, noise_specs = noise_specs,
                                individual_specs = individual_specs, Nrep = 100)
# Note that the individual specs override the general specs.

## Example specifying on whole groups (with helper function)
# We want 25\% noise on pillars in connectivity group,
# and 50\% noise on pillars in sustainability group.

# First find iCodes in those groups
p_conn <- get_iCodes_in_group(coin, "Conn", 2)
p_sust <- get_iCodes_in_group(coin, "Sust", 2)

# make list: the values first
individual_specs <- c(rep(0.25, length(p_conn)), rep(0.5, length(p_sust))) |>
  as.list()
# add the names
names(individual_specs) <- c(p_conn, p_sust)

# now run...
noisy_wts <- get_noisy_weights2(w = w_nom, noise_specs = noise_specs,
                                individual_specs = individual_specs, Nrep = 100)

}
\seealso{
\itemize{
\item \code{\link[=get_sensitivity]{get_sensitivity()}} Perform global sensitivity or uncertainty analysis on a COIN
}
}
