% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity.R
\name{get_perturbed_weight_samples}
\alias{get_perturbed_weight_samples}
\title{Generate samples of perturbed weights respecting requirement to sum to 1}
\usage{
get_perturbed_weight_samples(
  w,
  pert_by,
  Nrep,
  tolerance = 0.01,
  quietly = TRUE
)
}
\arguments{
\item{w}{Vector of numeric weights}

\item{pert_by}{A positive number representing the fraction by which to perturb weights.
E.g. setting to 0.5 will perturb weights +/- 50\% of their nominal values.}

\item{Nrep}{The number of weight sets to generate.}

\item{tolerance}{The tolerance (see description).}

\item{quietly}{Logical: if \code{FALSE} will also report the total number of weight
draws.}
}
\value{
A matrix with \code{Nrep} rows, where each row is a weight sample.
}
\description{
Given a vector of weights \code{w}, returns a matrix of \code{Nrep} rows where each row
is \code{w} with a random uniform perturbation applied, as specified by \code{pert_by}. Uses
a type of rejection sampling, with the requirement that weight sets must sum
to 1 +/- \code{tolerance}.
}
\examples{
# vector of four equal weights
w <- c(0.25, 0.25, 0.25, 0.25)

# perturb by +/-10\%, generate 10 vectors of perturbed weights
get_perturbed_weight_samples(w, pert_by = 0.1, Nrep = 10, quietly = FALSE)


}
