% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity.R
\name{get_sensitivity2}
\alias{get_sensitivity2}
\title{Sensitivity and uncertainty analysis of a coin (version 2)}
\usage{
get_sensitivity2(
  coin,
  SA_specs,
  N,
  SA_type = "UA",
  dset,
  iCode,
  Nboot = NULL,
  quietly = FALSE,
  check_addresses = TRUE,
  bump_geometric_mean = FALSE,
  report_progress = "bar",
  monitor_convergence = TRUE,
  converge_on = NULL
)
}
\arguments{
\item{coin}{A coin}

\item{SA_specs}{Specifications of the input uncertainties}

\item{N}{The number of regenerations}

\item{SA_type}{The type of analysis to run. \code{"UA"} runs an uncertainty analysis. \code{"SA"} runs a sensitivity
analysis (which anyway includes an uncertainty analysis).}

\item{dset}{The data set to extract the target variable from (passed to \code{\link[=get_data]{get_data()}}).}

\item{iCode}{The variable within \code{dset} to use as the target variable (passed to \code{\link[=get_data]{get_data()}}).}

\item{Nboot}{Number of bootstrap samples to take when estimating confidence intervals on sensitivity
indices.}

\item{quietly}{Set to \code{TRUE} to suppress progress messages.}

\item{check_addresses}{Logical: if \code{FALSE} skips the check of the validity of the parameter addresses. Default \code{TRUE},
but useful to set to \code{FALSE} if running this e.g. in a Rmd document (because may require user input).}

\item{bump_geometric_mean}{Logical: if \code{TRUE} will automatically deal with zero and negative values entering the geometric
mean by ensuring that the minimum value of each indicator is set to 1\% of its range. NOTE: this argument is currently
inactive until we decide if and how to implement this.}

\item{report_progress}{Controls the console output if \code{quietly = FALSE}. Either set to \code{"bar"} to see a progress bar, or
\code{"text"} to see text with a little more detail (although the latter will fill up the console).}

\item{monitor_convergence}{Logical: if \code{TRUE}, at every fifth iteration an estimation of the estimation error is calculated
via bootstrapping. This is recorded in the output list.}

\item{converge_on}{Optional parameter which specifies convergence. If specified, when the bootstrapped error
is less than \code{converge_on}, further iterations will be cancelled. Set to e.g. 0.01 which roughly translates as requiring
1\% error or less in rank estimations.}
}
\value{
Sensitivity analysis results as a list, containing:
\itemize{
\item \code{.$Scores} a data frame with a row for each unit, and columns are the scores for each replication.
\item \code{.$Ranks} as \code{.$Scores} but for unit ranks
\item \code{.$RankStats} summary statistics for ranks of each unit
\item \code{.$Para} a list containing parameter values for each run
\item \code{.$Nominal} the nominal scores and ranks of each unit (i.e. from the original COIN)
\item \code{.$Sensitivity} (only if \code{SA_type = "SA"}) sensitivity indices for each parameter. Also confidence intervals if \code{Nboot}
was specified.
\item \code{.$est_err} vector of estimated error at every fifth iteration (only if \code{monitor_convergence = TRUE})
\item Some information on the time elapsed, average time, and the parameters perturbed.
\item Depending on the setting of \code{store_results}, may also contain a list of Methods or a list of COINs for each replication.
}
}
\description{
This function performs global sensitivity and uncertainty analysis of a coin. You must specify which
parameters of the coin to vary, and the alternatives/distributions for those parameters. This function is a
modified version of \code{\link[=get_sensitivity]{get_sensitivity()}} (branched Nov 2023). See there for the general documentation.
In the Details below we focus on the modifications that are different from the COINr version.
}
\details{
This version also includes the possibility to switch between text or progress bar output to the console using
the \code{report_progress} argument.
}
\examples{
# for examples, see `vignette("sensitivity")`
# (this is because package examples are run automatically and this function can
# take a few minutes to run at realistic settings)

}
