% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{get_stats}
\alias{get_stats}
\title{Column-wise statistics}
\usage{
get_stats(x, ...)
}
\arguments{
\item{x}{Object (data frame or coin)}

\item{...}{Further arguments to be passed to methods.}
}
\value{
A data frame of statistics for each column
}
\description{
Reports various statistics from a data frame or coin.
}
\examples{
#

}
