% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_and_get.R
\name{get_units}
\alias{get_units}
\title{Get units of indicators}
\usage{
get_units(coin, iCodes = NULL)
}
\arguments{
\item{coin}{A coin}

\item{iCodes}{Some iCodes}
}
\value{
Vector of names
}
\description{
Given iCodes, returns corresponding units if available.
}
\examples{
#

}
