% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coin_multivariate.R
\name{getCorr}
\alias{getCorr}
\title{Get different types of correlation matrices}
\usage{
getCorr(
  COIN,
  dset,
  icodes = NULL,
  aglevs = NULL,
  cortype = "pearson",
  pval = 0.05,
  withparent = TRUE,
  grouplev = NULL
)
}
\arguments{
\item{COIN}{The COIN object}

\item{dset}{The data set to treat}

\item{icodes}{An optional list of character vectors where the first entry specifies the indicator/agg
codes to correlate against the second entry (also a specification of ind/agg codes)}

\item{aglevs}{The aggregation levels to take the two groups of indicators from. See \code{getIn()} for details.}

\item{cortype}{The type of correlation to calculate, either "pearson", "spearman", or "kendall".}

\item{pval}{The significance level for including correlations. Correlations with p < pval will be shown,
otherwise they will be plotted as white squares. Set to 0 to disable this.}

\item{withparent}{If \code{TRUE}, and \code{aglev[1] != aglev[2]}, will only return correlations of each row with its parent.}

\item{grouplev}{The aggregation level to group correlations by if \code{aglev[1] == aglev[2]}. By default, groups correlations into the
aggregation level above. Set to 0 to disable grouping and return the full matrix.}
}
\value{
A treated data set plus information about how the data was treated.
}
\description{
Helper function for plotting correlations between indicators. This retrieves subsets of correlation
matrices between different aggregation levels, in different formats.
}
