% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel.R
\name{getPackageBPPARAM}
\alias{getPackageBPPARAM}
\title{Get the default BiocParallel backend for this package}
\usage{
getPackageBPPARAM()
}
\value{
A BiocParallelParam object.
}
\description{
Get the default BiocParallel backend for this package
}
