% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coin_preanalyse.R
\name{getStats}
\alias{getStats}
\title{Get table of indicator statistics for any data set}
\usage{
getStats(
  COINobj,
  icodes = NULL,
  dset = "Raw",
  out2 = "COIN",
  cortype = "pearson",
  t_skew = 2,
  t_kurt = 3.5,
  t_colin = 0.9,
  t_denom = 0.7,
  t_missing = 65,
  IQR_coef = 1.5
)
}
\arguments{
\item{COINobj}{A list of indicator data, stuctured using the COIN_assemble function}

\item{icodes}{A character vector of indicator names to analyse. Defaults to all indicators.}

\item{dset}{The data set to analyse}

\item{out2}{Where to output the results: if "COIN" (default), outputs to the COIN, otherwise if "list", outputs to a separate list.}

\item{cortype}{The type of correlation to calculate, either "pearson", "spearman", or "kendall".}

\item{t_skew}{Skewness threshold}

\item{t_kurt}{Kurtosis threshold}

\item{t_colin}{Collinearity threshold (absolute value of correlation)}

\item{t_denom}{High correlation with denominator threshold}

\item{t_missing}{Missing data threshold, in percent}

\item{IQR_coef}{Interquartile range coefficient, used for identifying outliers}
}
\value{
Either an updated COIN object with relevant tables, or a list.
}
\description{
Takes a COIN , or data frame and returns a table of statistics, including max, min, median, mean, std, kurtosis, etc.
Flags indicators with possible outliers, and checks for collinearity with other indicators and denominators (if any).
Also checks number of unique values and percentage of zeros. Also returns correlation matrices and a table of outliers,
as a list.
}
\examples{
\dontrun{
# build ASEM COIN
ASEM <- assemble(IndData = ASEMIndData, IndMeta = ASEMIndMeta, AggMeta = ASEMAggMeta)
# get list of stats from raw data set
stat_list <- getStats(ASEM, dset = "Raw", out2 = "list")}

}
