% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coin_resultstable.R
\name{getStrengthNWeak}
\alias{getStrengthNWeak}
\title{Generate strengths and weaknesses for a specified unit}
\usage{
getStrengthNWeak(COIN, usel = NULL, topN = 5, bottomN = 5, withcodes = TRUE)
}
\arguments{
\item{COIN}{A COIN}

\item{usel}{A selected unit code}

\item{topN}{The top N indicators to report}

\item{bottomN}{The bottom N indicators to report}

\item{withcodes}{If TRUE (default), also includes a column of indicator codes. Setting to FALSE may be more useful
in generating reports, where codes are not helpful.}
}
\value{
A data frame
}
\description{
Generates a table of strengths and weaknesses for a selected unit, based on ranks. This currently only works
at the indicator level. Indicators with NAs are ignored.
}
\examples{
\dontrun{"test")}

}
