% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coin_resultstable.R
\name{getUnitSummary}
\alias{getUnitSummary}
\title{Generate unit summary table}
\usage{
getUnitSummary(COIN, usel, aglevs)
}
\arguments{
\item{COIN}{A COIN}

\item{usel}{A selected unit code}

\item{aglevs}{The aggregation levels to display results from.}
}
\value{
A data frame summary table.
}
\description{
Generates a summary table for a single unit. This is mostly useful in unit reports.
}
\examples{
\dontrun{"test")}

}
