% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treat.R
\name{GIIlog}
\alias{GIIlog}
\title{Log-transform a vector}
\usage{
GIIlog(x, na.rm = FALSE)
}
\arguments{
\item{x}{A numeric vector.}

\item{na.rm}{Set \code{TRUE} to remove \code{NA} values, otherwise returns \code{NA}.}
}
\value{
A log-transformed vector of data.
}
\description{
Performs a log transform on a numeric vector.
}
\details{
Specifically, this performs a "GII log" transform, which is what was encoded in the GII2020 spreadsheet.
}
\examples{
x <- runif(20)
GIIlog(x)

}
