% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{\%nin\%}
\alias{\%nin\%}
\title{Not in operator}
\usage{
x \%nin\% y
}
\arguments{
\item{x}{A scalar or vector}

\item{y}{A scalar or vector}
}
\value{
TRUE if x is not in y, FALSE otherwise
}
\description{
For convenience, rather than always \verb{!(x, \%in\% y)}
}
