% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coin_rew8r.R
\name{hicorrSP}
\alias{hicorrSP}
\title{Highly correlated indicators in same sub pillar}
\usage{
hicorrSP(COIN, dset = "Normalised", hicorval = 0.9, cortype = "pearson")
}
\arguments{
\item{COIN}{Data frame with first col indicator codes, second is weights, third is correlations}

\item{dset}{The data set to use for correlations.}

\item{hicorval}{A threshold to flag high correlation. Default 0.9.}

\item{cortype}{The type of correlation, either "pearson" (default), "spearman" or "kendall"}
}
\value{
A data frame with one entry for every indicator pair that is highly correlated within a sub-pillar.
Pairs are only reported once, i.e. only uses the upper triangle of the correlation matrix.
}
\description{
This returns a data frame of any highly correlated indicators within the same sub-pillar.
}
