% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{icodes_to_inames}
\alias{icodes_to_inames}
\title{Convert iCodes to iNames}
\usage{
icodes_to_inames(coin, iCodes)
}
\arguments{
\item{coin}{A coin}

\item{iCodes}{A vector of iCodes}
}
\value{
Vector of iNames
}
\description{
Convert iCodes to iNames
}
