% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{Impute.purse}
\alias{Impute.purse}
\title{Impute indicators}
\usage{
\method{Impute}{purse}(
  x,
  dset = NULL,
  f_i = NULL,
  f_i_para = NULL,
  impute_by = "column",
  group_level = NULL,
  normalise_first = NULL,
  write_to = NULL
)
}
\arguments{
\item{x}{A purse object}

\item{dset}{The data set to impute in each coin}

\item{f_i}{xx}

\item{f_i_para}{xx}

\item{impute_by}{xx}

\item{group_level}{xx}

\item{normalise_first}{xx}

\item{write_to}{xx}
}
\value{
An updated purse with imputed data sets added to each coin.
}
\description{
Impute indicators
}
\examples{
#
}
