% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{impute2.coin}
\alias{impute2.coin}
\title{Impute a data set in a coin}
\usage{
\method{impute2}{coin}(
  x,
  dset = NULL,
  f_i = NULL,
  f_i_para = NULL,
  impute_by = "column",
  group_level = NULL,
  norm_specs = NULL,
  ...,
  out2 = "coin",
  write_to = NULL
)
}
\arguments{
\item{x}{A coin class object}

\item{dset}{A data set within the coin}

\item{f_i}{An imputation function}

\item{f_i_para}{Further arguments to pass to \code{f_i}}

\item{impute_by}{s}

\item{group_level}{xx}

\item{norm_specs}{xx}

\item{...}{further arguments to be passed to \code{\link[=normalise]{normalise()}}}

\item{out2}{Either \code{"coin"} to return normalised data set back to the coin, or \code{df} to simply return a data
frame.}

\item{write_to}{Optional character string for naming the data set in the coin. Data will be written to
\code{.$Data[[write_to]]}. Default is \code{write_to == "Imputed"}.}
}
\value{

}
\description{
Impute a data set in a coin
}
\examples{
#
}
