% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{impute2.data.frame}
\alias{impute2.data.frame}
\title{Impute a data frame}
\usage{
\method{impute2}{data.frame}(x, f_i = NULL, f_i_para = NULL, impute_by = "column")
}
\arguments{
\item{x}{A data frame with only numeric columns.}

\item{f_i}{A function to use for imputation. By default, imputation is performed by simply substituting
the mean of non-\code{NA} values for each column at a time.}

\item{f_i_para}{Any additional parameters to pass to \code{f_i}, apart from \code{x}}

\item{impute_by}{Specifies how to impute: if \code{"column"}, passes each column separately as a numerical
vector to \code{f_i}; if \code{"row"}, passes each \emph{row} separately; and if \code{"df"} passes the entire data frame to
\code{f_i}. The function called by \code{f_i} should be compatible with the type of data passed to it.}
}
\value{

}
\description{
Impute a data frame using any function, either columnwise, rowwise or by the whole data frame in one
shot.
}
\details{
This function only accepts data frames with all numeric columns. It imputes any NAs in the data frame
by invoking the function \code{f_i} and any optional arguments \code{f_i_para} on each column at a time (if
\code{impute_by = "column"}), or on each row at a time (if \code{impute_by = "row"}), or by passing the entire
data frame to \code{f_i} if \code{impute_by = "df"}.

Clearly, the function \code{f_i} needs to be able to accept with the data class passed to it - if
\code{impute_by} is \code{"row"} or \code{"column"} this will be a numeric vector, or if \code{"df"} it will be a data
frame. Checks are made on the format of the data returned by imputation functions, to ensure the
type and that non-\code{NA} values have not been inadvertently altered.
}
\examples{
#
}
