% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{impute2}
\alias{impute2}
\title{Impute indicator data sets}
\usage{
impute2(x, ...)
}
\arguments{
\item{x}{Object to be imputed}

\item{...}{Further arguments to be passed to methods.}
}
\value{
An object of the same class as \code{x}, but imputed.
}
\description{
Impute indicator data sets
}
\examples{
#

}
