% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coin_indicatordash.R
\name{indDash}
\alias{indDash}
\title{Indicator visualisation dashboard}
\usage{
indDash(COIN, icodes = NULL, dset = "Raw")
}
\arguments{
\item{COIN}{The COIN object, or a data frame of indicator data.}

\item{icodes}{A set of indicator codes to include. Defaults to all indicators.}

\item{dset}{The data set to plot.}
}
\value{
Interactive visualisation
}
\description{
Generates an interactive visualisation of one or two indicators at a time. Requires Shiny and an active R session.
}
\examples{
\dontrun{indDash(COIN, icodes = NULL, dset = "raw")}

}
