% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coin_resultsdash.R
\name{iplot_bar}
\alias{iplot_bar}
\title{Bar chart}
\usage{
iplot_bar(COINobj, dset = "Raw", isel)
}
\arguments{
\item{COINobj}{The COIN object, or a data frame of indicator data.}

\item{dset}{The data set to plot.}

\item{isel}{The selected indicator code or aggregate}
}
\value{
Interactive bar chart
}
\description{
Generates an interactive bar chart. This function is simply a wrapper for
the Plotly bar chart function.
}
\examples{
\dontrun{coin_indicatordash(COINobj, inames = NULL, dset = "raw")}

}
