% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coin_resultsdash.R
\name{iplot_map}
\alias{iplot_map}
\title{Choropleth map}
\usage{
iplot_map(COINobj, dset = "Raw", isel)
}
\arguments{
\item{COINobj}{The COIN object, or a data frame of indicator data.}

\item{dset}{The data set to plot.}

\item{isel}{The selected indicator code or aggregate}
}
\value{
Interactive map
}
\description{
Generates an interactive choropleth map of specified indicator data. Only works on national
level data (i.e. one point per country), with countries labeled by ISO alpha-3 codes. See e.g.
https://www.iso.org/iso-3166-country-codes.html. This function is simply a wrapper for
the Plotly choropleth map function.
}
\examples{
\dontrun{coin_indicatordash(COINobj, inames = NULL, dset = "raw")}

}
