% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coin_resultsdash.R
\name{iplot_radar}
\alias{iplot_radar}
\title{Radar chart}
\usage{
iplot_radar(COINobj, dset = "Raw", levsel = NULL, usel = NULL)
}
\arguments{
\item{COINobj}{The COIN object, or a data frame of indicator data.}

\item{dset}{The data set to use in the table}

\item{levsel}{The selected aggregation level to take indicator data from,
where 1 is the base indicator level, and 2, 3 etc are higher aggregation levels}
}
\value{
Interactive table
}
\description{
Generates an interactive radar chart for a specified unit.
}
\examples{
\dontrun{coin_indicatordash(COINobj, inames = NULL, dset = "raw")}

}
