% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coin_resultsdash.R
\name{iplot_table}
\alias{iplot_table}
\title{Results table}
\usage{
iplot_table(COINobj, dset = "Raw", isel = NULL)
}
\arguments{
\item{COINobj}{The COIN object, or a data frame of indicator data.}

\item{dset}{The data set to use in the table}

\item{isel}{The selected indicator codes (default all in dset)}
}
\value{
Interactive table
}
\description{
Generates an interactive table of data. For use in e.g. Shiny or html documents.
}
\examples{
\dontrun{coin_indicatordash(COINobj, inames = NULL, dset = "raw")}

}
