% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coin_resultsdash.R
\name{iplotRadar}
\alias{iplotRadar}
\title{Radar chart}
\usage{
iplotRadar(COIN, dset = "Raw", usel = NULL, aglev = NULL, isel = NULL)
}
\arguments{
\item{COIN}{The COIN object, or a data frame of indicator data.}

\item{dset}{The data set to use in the table}

\item{usel}{Character vector of unit code(s) to plot data from}

\item{aglev}{The selected aggregation level to take indicator data from,
where 1 is the base indicator level, and 2, 3 etc are higher aggregation levels}

\item{isel}{The indicator or aggregation code(s) to plot}
}
\value{
Interactive table
}
\description{
Generates an interactive radar chart for a specified unit.
}
\examples{
\dontrun{
iplotRadar(ASEM, dset = "Aggregated", usel = c("AUT", "CHN"), isel = "Physical", aglev = 1)
}

}
