% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coin_resultsdash.R
\name{iplotTable}
\alias{iplotTable}
\title{Results table}
\usage{
iplotTable(
  COIN,
  dset = "Raw",
  isel = NULL,
  aglev = NULL,
  nround = 1,
  extra_cols = FALSE
)
}
\arguments{
\item{COIN}{The COIN object, or a data frame of indicator data.}

\item{dset}{The data set to use in the table}

\item{isel}{The selected indicator codes (default all in dset)}

\item{aglev}{The aggregation level to select data from}

\item{nround}{The number of decimal places to round scores to (default 1).}

\item{extra_cols}{If FALSE (default), excludes group columns and similar. If TRUE, includes them.}
}
\value{
Interactive table
}
\description{
Generates an interactive table of data. For use in e.g. Shiny or html documents.
}
\examples{
\dontrun{iplotTable(COIN, dset = "Aggregated", isel = NULL)}

}
