% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_added.R
\name{is.COIN}
\alias{is.COIN}
\title{Check if an object is a COIN}
\usage{
is.COIN(obj)
}
\arguments{
\item{x}{An input object to test}
}
\value{
Logical: \code{TRUE} if input is a COIN, otherwise \code{FALSE}
}
\description{
Returns \code{TRUE} if an input object is a COIN, otherwise \code{FALSE} if not. This checks for the older
"COIN" class, which was supported up to COINr v0.6.1. Later versions use the new "coin" class.
}
\examples{
# something that is not a COIN
is.COIN("hello")

}
