% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_added.R
\name{is.purse}
\alias{is.purse}
\title{Check if object is purse class}
\usage{
is.purse(x)
}
\arguments{
\item{x}{An object to be checked.}
}
\value{
Logical
}
\description{
Check if object is purse class
}
