% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treat.R
\name{kurt}
\alias{kurt}
\title{Calculate kurtosis}
\usage{
kurt(x, na.rm = FALSE)
}
\arguments{
\item{x}{A numeric vector.}

\item{na.rm}{Set \code{TRUE} to remove \code{NA} values, otherwise returns \code{NA}.}
}
\value{
A kurtosis value (scalar).
}
\description{
Calculates kurtosis of the values of a numeric vector. This uses the same definition of kurtosis as
as the "kurtosis()" function in the e1071 package, where \code{type == 2}, which is equivalent to the definition of kurtosis used in Excel.
}
\examples{
x <- runif(20)
kurt(x)

}
