% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{lengthen}
\alias{lengthen}
\title{Make correlation matrix long}
\usage{
lengthen(X, cols = NULL)
}
\arguments{
\item{X}{a square correlation matrix}
}
\value{
A long format data frame
}
\description{
Only for correlation matrices: make long to avoid reshape2 package or similar.
}
