% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coin_treat.R
\name{loggish}
\alias{loggish}
\title{Log-type transformation of a vector}
\usage{
loggish(x, ltype, params)
}
\arguments{
\item{x}{A vector or column of data to transform}

\item{ltype}{The type of log transformation}

\item{params}{Some extra params to pass}
}
\description{
This applies various simple transformations, to be used by the treat() function.
}
\seealso{
\itemize{
\item \code{\link[=treat]{treat()}} Outlier treatment
}
}
