% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalise.R
\name{n_borda}
\alias{n_borda}
\title{Normalise using Borda scores}
\usage{
n_borda(x, ties.method = "min")
}
\arguments{
\item{x}{A numeric vector}

\item{ties.method}{This argument is passed to \code{\link[base:rank]{base::rank()}} - see there for details.}
}
\value{
Numeric vector
}
\description{
Calculates Borda scores as rank(x) - 1.
}
\examples{
x <- runif(20)
n_borda(x)

}
