% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalise.R
\name{n_fracmax}
\alias{n_fracmax}
\title{Normalise as fraction of max value}
\usage{
n_fracmax(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
Numeric vector
}
\description{
The ratio of each value of \code{x} to \code{max(x)}.
}
\details{
$ x / x_{max} $
}
\examples{
x <- runif(20)
n_fracmax(x)

}
