% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalise.R
\name{n_prank}
\alias{n_prank}
\title{Normalise using percentile ranks}
\usage{
n_prank(x, ties.method = "min")
}
\arguments{
\item{x}{A numeric vector}

\item{ties.method}{This argument is passed to \code{\link[base:rank]{base::rank()}} - see there for details.}
}
\value{
Numeric vector
}
\description{
Calculates percentile ranks of a numeric vector using "sport" ranking. Ranks are calculated by \code{\link[base:rank]{base::rank()}}
and converted to percentile ranks. The \code{ties.method} can be changed - this is directly passed to
\code{\link[base:rank]{base::rank()}}.
}
\examples{
x <- runif(20)
n_prank(x)

}
