% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalise.R
\name{n_rank}
\alias{n_rank}
\title{Normalise using ranks}
\usage{
n_rank(x, ties.method = "min")
}
\arguments{
\item{x}{A numeric vector}

\item{ties.method}{This argument is passed to \code{\link[base:rank]{base::rank()}} - see there for details.}
}
\value{
Numeric vector
}
\description{
This is simply a wrapper for \code{\link[base:rank]{base::rank()}}. Higher scores will give higher ranks.
}
\examples{
x <- runif(20)
n_rank(x)

}
