% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_coin.R
\name{new_coin}
\alias{new_coin}
\title{Create a new coin}
\usage{
new_coin(iData, iMeta)
}
\arguments{
\item{iData}{The indicator data and metadata of each unit}

\item{iMeta}{Indicator metadata}
}
\value{
A "coin" object
}
\description{
Creates a new coin class object.
}
\examples{
#

}
\seealso{

}
