% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unbalanced_coin.R
\name{new_unbalanced_coin}
\alias{new_unbalanced_coin}
\title{Create a new unbalanced coin}
\usage{
new_unbalanced_coin(
  iData,
  iMeta,
  exclude = NULL,
  split_to = NULL,
  level_names = NULL,
  retain_all_uCodes_on_split = FALSE,
  quietly = FALSE
)
}
\arguments{
\item{iData}{The indicator data and metadata of each unit}

\item{iMeta}{Indicator metadata}

\item{exclude}{Optional character vector of any indicator codes (\code{iCode}s) to exclude from the coin(s).}

\item{split_to}{This is used to split panel data into multiple coins, a so-called "purse". Should be either
\code{"all"}, or a subset of entries in \code{iData$Time}. See Details.}

\item{level_names}{Optional character vector of names of levels. Must have length equal to the number of
levels in the hierarchy (\code{max(iMeta$Level, na.rm = TRUE)}).}

\item{retain_all_uCodes_on_split}{Logical: if panel data is input and split to a purse using \code{split_to}, this controls how
units with no data at certain time points are handled. If set \code{FALSE}, then unit at time t with no data in any indicators
will be removed completely from the coin for that time point. If \code{TRUE}, all units will be included in every time point. The latter option
may be useful if you impute over time.}

\item{quietly}{If \code{TRUE}, suppresses all messages}
}
\value{
An object of class \code{unbalanced_coin} that delegates to the standard coin pipeline
while keeping track of placeholder nodes used to balance the hierarchy internally.
}
\description{
Create a new unbalanced coin
}
