% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coin_sensitivity.R
\name{noisyWeights}
\alias{noisyWeights}
\title{Noisy replications of weights}
\usage{
noisyWeights(w, noise_specs, Nrep)
}
\arguments{
\item{w}{A data frame of weights, in the format found in .$Parameters$Weights}

\item{noise_specs}{a data frame with columns:
\itemize{
\item AgLevel: The aggregation level to apply noise to
\item NoiseFactor. The size of the perturbation: setting e.g. 0.2 peturbs by +/- 20\% of nominal values
}}

\item{Nrep}{The number of weight replications to generate.}
}
\value{
A list of sets of weights (data frames)
}
\description{
Given a set of weights, this function returns multiple replicates of the weights, with added
noise. This is intended for use in uncertainty and sensitivity analysis.
}
\details{
Weights are expected to be in a long data frame format with columns Aglevel, Code and Weight, as
used inside COINs.

Noise is added using the noise_specs argument, which is specified by a data frame with columns
AgLevel and NoiseFactor. The aggregation level refers to number of the aggregation level to target
while the NoiseFactor refers to the size of the perturbation. If e.g. a row is AgLevel = 1 and
NoiseFactor = 0.2, this will allow the weights in aggregation level 1 to deviate by +/- 20\% of their
nominal values (the values in w).
}
