% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalise.R
\name{Normalise.purse}
\alias{Normalise.purse}
\title{Create normalised data set in a purse of coins}
\usage{
\method{Normalise}{purse}(
  x,
  dset = NULL,
  default_specs = NULL,
  indiv_specs = NULL,
  directions = NULL,
  global = TRUE,
  write_to = NULL
)
}
\arguments{
\item{x}{A purse object}

\item{dset}{The data set to normalise in each coin}

\item{default_specs}{Default specifications}

\item{indiv_specs}{Individual specifications}

\item{directions}{If \code{NULL}, extracts directions from indicator metadata, i.e. the \code{iMeta} data frame
that was passed to \code{\link[=new_coin]{new_coin()}}. Else \code{directions} should be a vector with entries either -1 or 1, in
order of the columns of the data set.}
}
\value{
An updated purse with new normalised data sets added at \code{.$Data$Normalised} in each coin
}
\description{
Normalisation is a special case in terms of operating on a purse of coins. This is because, when
dealing with time series data, it is often desirable to normalise over the whole panel data set
rather than independently for each time point. This makes the resulting index and aggregates comparable
over time.
}
\details{
The same specifications are passed to each coin in the purse. This means that each coin is normalised
using the same set of specifications and directions. If you need control over individual coins, you
will have to normalise coins individually.
}
\examples{
#
}
