% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalise.R
\name{normalise2.data.frame}
\alias{normalise2.data.frame}
\title{Normalise a data frame}
\usage{
\method{normalise2}{data.frame}(x, default_specs = NULL, indiv_specs = NULL, directions = NULL)
}
\arguments{
\item{x}{A data frame}

\item{default_specs}{Specifications to apply to all columns, apart from those specified by \code{indiv_specs}.}

\item{indiv_specs}{Specifications applied to specific columns, overriding those specified in \code{default_specs}}

\item{directions}{An optional data frame containing the following columns:
\itemize{
\item \code{iCode} The indicator code, corresponding to the column names of the data frame
\item \code{Direction} numeric vector with entries either \code{-1} or \code{1}
If \code{directions} is not specified, the directions will all be assigned as \code{1}. Non-numeric columns do not need
to have directions assigned.
}}
}
\value{

}
\description{
Normalise a data frame
}
