% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalise.R
\name{normalise2}
\alias{normalise2}
\title{Normalise indicator data sets}
\usage{
normalise2(x, ...)
}
\arguments{
\item{x}{Object to be normalised}

\item{...}{Further arguments to be passed to methods.}
}
\value{
An updated GII with normalised data set added
}
\description{
Normalisation is specified using the \code{ntype} and \code{npara} arguments. In these, \code{ntype} should be a character
string which is the name of a normalisation
function. For example, \code{ntype = "minmax"} calls the \code{\link[=minmax]{minmax()}} function. \code{npara} is any
further argument to \code{ntype}. This means that any function can be passed to \code{\link[=normalise]{normalise()}}, as long as it
is of the form \code{f(x, npara)}, where \code{x} is a vector (i.e. one column of data) and should return a
similar vector of normalised data. See \code{\link[=minmax]{minmax()}} for an example. If a function has more than one
additional argument, then \code{npara} can be passed as a list and a wrapper could be created.
}
\examples{
#

}
