% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity.R
\name{plot_convergence}
\alias{plot_convergence}
\title{Convergence plot of SA results}
\usage{
plot_convergence(SA_res)
}
\arguments{
\item{SA_res}{Output list from \code{\link[=get_sensitivity2]{get_sensitivity2()}}}
}
\value{
A plot
}
\description{
Quick plot of the convergence of the SA results. Only works if \code{monitor_convergence = TRUE}
was specified when running \code{\link[=get_sensitivity2]{get_sensitivity2()}}.
}
\examples{
#
}
