% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coin_indplot.R
\name{plotIndDot}
\alias{plotIndDot}
\title{Dot plots of single indicator with highlighting}
\usage{
plotIndDot(
  COIN,
  dset = NULL,
  icode = NULL,
  usel = NULL,
  marker_type = "circle",
  add_stat = NULL,
  show_ticks = TRUE,
  plabel = NULL,
  usel_code = TRUE
)
}
\arguments{
\item{COIN}{The COIN}

\item{dset}{The source data set to use for indicator data}

\item{icode}{An indicator code to plot.}

\item{usel}{A unit or set of units (as a string or character vector) to highlight.}

\item{marker_type}{The type of marker, either \code{"circle"} (default) or \code{"cross"}, or a marker number to pass to ggplot2 (0-25).}

\item{add_stat}{A statistic to overlay, either \code{"mean"}, \code{"median"} or else a specified value.}

\item{plabel}{Controls the labelling of the indicator. If not specified, returns the indicator name,
plus units if found. Otherwise if \code{"indname"}, returns only indicator name, if \code{"indname+unit"}, returns
indicator name plus unit (if found), if \code{"unit"} returns only unit (if found), otherwise if \code{"none"},
displays no text. Finally, any other string can be passed, so e.g. \code{"My indicator"} will display this on the
axis.}

\item{usel_label}{If \code{TRUE} (default) also labels selected units with their unit codes. \code{FALSE} to disable.}
}
\value{
Plots generated with \strong{ggplot2}. These can be edited further with \strong{ggplot2} commands.
}
\description{
Plots a single indicator as a line of dots, and optionally highlights a selected unit.
}
\examples{
# add

}
