% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coin_sensitivity.R
\name{plotSA}
\alias{plotSA}
\title{Plot sensitivity indices}
\usage{
plotSA(SAresults, ptype = "bar")
}
\arguments{
\item{SAresults}{A list of sensitivity/uncertainty analysis results from COINr::sensitivity().}

\item{ptype}{Type of plot to generate - either "bar", "pie" or "box"}
}
\description{
Plots sensitivity indices as bar or pie charts.
}
\details{
To use this function you first need to run sensitivity(). Then enter the resulting list as the
\code{SAresults} argument here. I haven't provided an example because of the time required for performing
a sensitivity analysis.
See \href{https://bluefoxr.github.io/COINrDoc/sensitivity-analysis.html}{COINr online documentation} for more details.
}
\seealso{
\itemize{
\item \code{\link[=sensitivity]{sensitivity()}} Perform global sensitivity or uncertainty analysis on a COIN
\item \code{\link[=plotSARanks]{plotSARanks()}} Plot confidence intervals on ranks following a sensitivity analysis
}
}
