% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coin_sensitivity.R
\name{plotSARanks}
\alias{plotSARanks}
\title{Plot ranks from an uncertainty/sensitivity analysis}
\usage{
plotSARanks(SAresults, plot_units = NULL, order_by = "nominal")
}
\arguments{
\item{SAresults}{A list of sensitivity/uncertainty analysis results from COINr::sensitivity().}

\item{plot_units}{A character vector of units to plot. Defaults to all units. You can also set
to "top10" to only plot top 10 units, and "bottom10" for bottom ten.}

\item{order_by}{If set to "nominal", orders the rank plot by nominal ranks
(i.e. the original ranks prior to the sensitivity analysis). Otherwise if "median", orders by
median ranks.}
}
\description{
Plots the ranks resulting from an uncertainty and sensitivity analysis, in particular plots
the median, and 5th/95th percentiles of ranks.
}
\details{
To use this function you first need to run sensitivity().
}
