% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_added.R
\name{print.coin}
\alias{print.coin}
\title{Print coin}
\usage{
\method{print}{coin}(x, ...)
}
\arguments{
\item{x}{A coin}

\item{...}{Arguments to be passed to or from other methods.}
}
\value{
Text output
}
\description{
Some details about the coin
}
