% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_added.R
\name{print.purse}
\alias{print.purse}
\title{Print purse}
\usage{
\method{print}{purse}(x, ...)
}
\arguments{
\item{x}{A purse}

\item{...}{Arguments to be passed to or from other methods.}
}
\value{
Text output
}
\description{
Some details about the purse
}
