% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qfuncs.R
\name{qNormalise.data.frame}
\alias{qNormalise.data.frame}
\title{Quick normalisation of a data frame}
\usage{
\method{qNormalise}{data.frame}(
  x,
  f_n = "n_minmax",
  f_n_para = list(l_u = c(0, 100)),
  directions = NULL,
  ...
)
}
\arguments{
\item{x}{A numeric data frame}

\item{f_n}{Name of a normalisation function (as a string) to apply to each column of \code{x}. Default \code{"n_minmax"}.}

\item{f_n_para}{Any further arguments to pass to \code{f_n}, as a named list.}

\item{directions}{An optional data frame containing the following columns:
\itemize{
\item \code{iCode} The indicator code, corresponding to the column names of the data frame
\item \code{Direction} numeric vector with entries either \code{-1} or \code{1}
If \code{directions} is not specified, the directions will all be assigned as \code{1}. Non-numeric columns do not need
to have directions assigned.
}}

\item{...}{arguments passed to or from other methods.}
}
\value{
A data frame
}
\description{
Quick normalisation of a data frame
}
\examples{
#
}
