% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qfuncs.R
\name{qNormalise}
\alias{qNormalise}
\title{Quick normalisation}
\usage{
qNormalise(x, ...)
}
\arguments{
\item{x}{Object to be normalised}

\item{...}{arguments passed to or from other methods.}
}
\value{
A normalised object
}
\description{
This is a generic wrapper function for \code{\link[=Normalise]{Normalise()}}, which offers a simpler syntax but less flexibility.
}
\details{
See individual method documentation:
\itemize{
\item \code{\link[=qNormalise.data.frame]{qNormalise.data.frame()}}
\item \code{\link[=qNormalise.coin]{qNormalise.coin()}}
\item \code{\link[=qNormalise.unbalanced_coin]{qNormalise.unbalanced_coin()}}
\item \code{\link[=qNormalise.purse]{qNormalise.purse()}}
}
}
