% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coin_resultstable.R
\name{rankDF}
\alias{rankDF}
\title{Convert a data frame to ranks}
\usage{
rankDF(df)
}
\arguments{
\item{df}{A data frame}
}
\value{
Data frame with ranks
}
\description{
Replaces all numerical columns of a data frame with their ranks. Uses sport ranking, i.e. ties
share the highest rank place.
}
\examples{
\dontrun{"test")}

}
