% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coin_regen.R
\name{regen}
\alias{regen}
\title{Regenerate COIN object}
\usage{
regen(COINold, quietly = FALSE)
}
\arguments{
\item{COINold}{COIN object containing specs on how to regenerate}

\item{quietly}{Logical: if TRUE suppresses all messages from COINr functions (warnings may still occur though)}
}
\value{
An updated COIN object, recalculated
}
\description{
Function to regenerate the results of the COIN, using the methodological parameters stored in
.$Method. This function calls the construction functions of COINr in the order that they are
found in .$Method, along with any custom code found in .$Method$Custom.
}
