% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity.R
\name{regen_edit}
\alias{regen_edit}
\title{Regenerate an edited coin}
\usage{
regen_edit(l_para, addresses, coin)
}
\arguments{
\item{l_para}{A list of parameter values to change. Should be of the format \code{list(para_name = new_value)}, where
\code{new_value}}

\item{addresses}{A list or character vector of addresses. \code{names(addresses)} must correspond to \code{names(l_para)}.}

\item{coin}{A coin, to be edited.}
}
\value{
A regenerated coin
}
\description{
This is similar to \code{\link[=edit_coin]{edit_coin()}} but works with a list of parameters to change, rather than one, and also outputs
a regenerated coin.
}
\examples{
#
}
