% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DFtools.R
\name{replace_df}
\alias{replace_df}
\title{Replace multiple values in a data frame}
\usage{
replace_df(df, lookup)
}
\arguments{
\item{df}{A data frame or a vector}

\item{lookup}{A data frame with columns \code{old} (the values to be replaced) and \code{new} the values to replace with. See details.}
}
\value{
A data frame with replaced values
}
\description{
Given a data frame (or vector), this function replaces values according to a look up table or dictionary. In COINr this may
be useful for exchanging categorical data with numeric scores, prior to assembly. Or for changing codes.
}
\details{
The lookup data frame must not have any duplicated values in the \code{old} column. This function looks for exact matches of
elements of the \code{old} column and replaces them with the corresponding value in the \code{new} column. For each row of \code{lookup},
the class of the old value must match the class of the new value. This is to keep classes of data frames columns consistent.
If you wish to replace with a different class, you should convert classes in your data frame before using this function.

This function replaces the now-defunct \code{replaceDF()} from COINr < v1.0.
}
\examples{
# replace sub-pillar codes in ASEM indicator metadata
codeswap <- data.frame(old = c("Conn", "Sust"), new = c("SI1", "SI2"))
# swap codes in both iMeta
replace_df(ASEM_iMeta, codeswap)

}
