% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coin_resultsdash.R
\name{resultsDash}
\alias{resultsDash}
\title{Results visualisation dashboard}
\usage{
resultsDash(COIN, dset = "Aggregated")
}
\arguments{
\item{COIN}{The COIN object, or a data frame of indicator data.}

\item{dset}{The data set to plot.}
}
\value{
Interactive visualisation
}
\description{
Generates an interactive visualisation of results. Requires Shiny and an active R session.
}
\examples{
\dontrun{coin_indicatordash(COIN, icodes = NULL, dset = "raw")}

}
