% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coin_rew8r.R
\name{rew8r}
\alias{rew8r}
\title{Reweight indicators}
\usage{
rew8r(COIN)
}
\arguments{
\item{COIN}{COIN object}
}
\value{
An updated COIN object with additional sets of weights.
}
\description{
Interactive gadget which lets you adjust weights and see the effects. Weights can be saved with new names to the COIN object.
}
