% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIn.R
\name{roundDF}
\alias{roundDF}
\title{Round down a data frame}
\usage{
roundDF(df, decimals = 2)
}
\arguments{
\item{df}{A data frame to input}

\item{decimals}{The number of decimal places to round to (default 2)}
}
\value{
A data frame, with any numeric columns rounded to the specified amount.
}
\description{
Tiny function just to round down a data frame for display in a table.
}
\examples{
roundDF( as.data.frame(matrix(runif(20),10,2)), decimals = 3)

}
\seealso{
\itemize{
\item \code{\link[=rankDF]{rankDF()}} Replace data frame numbers with ranks.
}
}
