% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity.R
\name{sample_2_para}
\alias{sample_2_para}
\title{Convert a numeric sample to parameter values}
\usage{
sample_2_para(x, distribution, dist_type = "discrete", checks = TRUE)
}
\arguments{
\item{x}{A numeric vector with values between 0 and 1}

\item{distribution}{The distribution to sample using \code{x} - see details.}

\item{dist_type}{Either \code{"discrete"} or \code{"continuous"} - see details.}

\item{checks}{Logical: if \code{TRUE} runs some checks on inputs, set to \code{FALSE} to increase speed slightly.}
}
\value{

}
\description{
Converts a numeric sample \code{x}, which should have values between 0 and 1, to a corresponding vector or list of
parameter values, based on \code{distribution}.
}
\details{
The \code{distribution} argument specifies how to map \code{x} to parameter values and can be used in two different ways,
depending on \code{dist_type}. If \code{dist_type = "discrete"}, then \code{distribution} should be a vector or list of alternative
parameter values (or objects). Each entry of \code{x} is mapped to an entry from \code{distribution} by treating \code{distribution}
as a discrete uniform distribution over its entries.

If \code{dist_type = "continuous"}, \code{distribution} is assumed to be a continuous uniform distribution, such that
\code{distribution} is a 2-length numeric vector with the first value being the lower bound, and the second value the
upper bound. For example, if \code{distribution = c(5, 10)}, then \code{x} will be mapped onto a continuous uniform distribution
bounded by 5 and 10.
}
\examples{
#
}
