% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screen_units.R
\name{Screen}
\alias{Screen}
\title{Screen units based on data availability}
\usage{
Screen(x, ...)
}
\arguments{
\item{x}{Object to be screened}

\item{...}{arguments passed to or from other methods.}
}
\value{
An object of the same class as \code{x}
}
\description{
This is a generic function for screening units/rows based on data availability. See method documentation
for more details:
}
\details{
This function replaces the now-defunct \code{checkData()} from COINr < v1.0.
\itemize{
\item \code{\link[=Screen.data.frame]{Screen.data.frame()}}
\item \code{\link[=Screen.coin]{Screen.coin()}}
\item \code{\link[=Screen.unbalanced_coin]{Screen.unbalanced_coin()}}
\item \code{\link[=Screen.purse]{Screen.purse()}}
}
}
