% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coin_sensitivity.R
\name{sensitivity}
\alias{sensitivity}
\title{Sensitivity analysis}
\usage{
sensitivity(
  COIN,
  v_targ,
  SA_specs,
  N,
  SA_type = "UA",
  NrepWeights = 1000,
  store_results = "results+params",
  Nboot = NULL,
  quietly = FALSE
)
}
\arguments{
\item{COIN}{A COIN (this function does not support data frame input)}

\item{v_targ}{The target variable to perform SA or UA on. Currently just supports one variable, which
should be an indicator/aggregate code present in .$Data$Aggregated.}

\item{SA_specs}{A list which specifies which variables to perturb, and which alternatives/distributions to use}

\item{N}{The number of Monte Carlo replications}

\item{SA_type}{The type of analysis to run. "UA" runs an uncertainty analysis.}

\item{NrepWeights}{The number of weight-replications to generate. Default 1000.}

\item{store_results}{Which results to store
\itemize{
\item "onlyresults" only stores scores, ranks and rank statistics (e.g. mean, median, quantiles)
\item "results+params" (default) stores all of the above, plus a record of the parameter values used for each replication
\item "results+method" stores all of the above, plus the full .$Method list of each replication
\item "results+COIN" stores all results and the complete COIN of each replication
}}

\item{Nboot}{Number of bootstrap draws for estimates of confidence intervals on sensitivity indices.
If this is not specified, bootstrapping is not applied. Ignored if SA_type = "UA".}

\item{quietly}{If FALSE (default), gives progress messages. Set TRUE to suppress these.}
}
\value{
Sensitivity analysis results
}
\description{
Performs global uncertainty and sensitivity analysis on a COIN.
}
\details{
To perform a sensitivity or uncertainty analysis, you must specify \emph{which} parameters/assumptions to vary and \emph{what} their alternative
values are. This is the \code{SA_specs} argument below. To understand how this works, please see the \href{https://bluefoxr.github.io/COINrDoc/sensitivity-analysis.html#sensitivity-in-coinr}{COINr online documentation}.
}
\examples{
\dontrun{

# build ASEM COIN up to aggregation
ASEM <- build_ASEM()

# define noise to be applied to weights
nspecs <- data.frame(AgLevel = c(2,3), NoiseFactor = c(0.25,0.25))

# create list specifying assumptions to vary and alternatives
SAspecs <- list(
  impute = list(imtype = c("indgroup_mean", "ind_mean", "none")),
  normalise = list(ntype = c("minmax", "rank", "dist2max")),
  weights = list(NoiseSpecs = nspecs, Nominal = "Original")
)

# run uncertainty analysis
SAresults <- sensitivity(ASEM, v_targ = "Index",
                         SA_specs = SAspecs,
                         N = 100,
                         SA_type = "UA")

# to run a sensitivity analysis set SA_type = "SA" (takes longer)}

}
\seealso{
\itemize{
\item \code{\link[=plotSARanks]{plotSARanks()}} Plot confidence intervals of ranks following UA or SA
\item \code{\link[=plotSA]{plotSA()}} Plot sensitivity indices following a sensitivity analysis
}
}
