% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_default}
\alias{set_default}
\title{Set default arg}
\usage{
set_default(x, x_default)
}
\arguments{
\item{x}{The argument}

\item{x_default}{The default to set}
}
\value{
the parameter
}
\description{
A shortcut
}
