% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel.R
\name{setPackageBPPARAM}
\alias{setPackageBPPARAM}
\title{Set the default BiocParallel backend for this package}
\usage{
setPackageBPPARAM(param)
}
\arguments{
\item{param}{A BiocParallelParam object.}
}
\value{
The previously set BiocParallelParam object (invisibly).
}
\description{
Set the default BiocParallel backend for this package
}
