% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DFtools.R
\name{signifDF}
\alias{signifDF}
\title{Round a data frame to specified significant figures}
\usage{
signifDF(df, digits = 3)
}
\arguments{
\item{df}{A data frame to input}

\item{digits}{The number of decimal places to round to (default 3)}
}
\value{
A data frame, with any numeric columns rounded to the specified amount.
}
\description{
Tiny function just to round down a data frame by significant figures for display in a table, ignoring non-numeric columns.
}
\examples{
signifDF( as.data.frame(matrix(runif(20),10,2)), digits = 3)

}
\seealso{
\itemize{
\item \code{\link[=rankDF]{rankDF()}} Replace data frame numbers with ranks.
}
}
