% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treat.R
\name{skew}
\alias{skew}
\title{Calculate skewness}
\usage{
skew(x, na.rm = FALSE)
}
\arguments{
\item{x}{A numeric vector.}

\item{na.rm}{Set \code{TRUE} to remove \code{NA} values, otherwise returns \code{NA}.}
}
\value{
A skewness value (scalar).
}
\description{
Calculates skewness of the values of a numeric vector. This uses the same definition of skewness as
\code{\link[e1071:skewness]{e1071::skewness()}} where \code{type == 2}, which is equivalent to the definition of skewness used in Excel.
}
\examples{
x <- runif(20)
skew(x)

}
