% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screen_units.R
\name{split_by_numeric}
\alias{split_by_numeric}
\title{Splits data frame into numeric and non-numeric columns}
\usage{
split_by_numeric(x)
}
\value{
A list with \code{.$not_numeric} containing a data frame with non-numeric columns, and \code{.$numeric} being
a data frame containing only numeric columns.
}
\description{
Splits data frame into numeric and non-numeric columns
}
\examples{
#
}
