% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_coin.R
\name{split_iData}
\alias{split_iData}
\title{Split iData}
\usage{
split_iData(iData, split_to)
}
\arguments{
\item{iData}{A data frame of indicator panel data}

\item{split_to}{Either \code{"all"} (one \code{iData} for each unique entry in \code{iData$Time}), or a vector containing a
subset of entries in \code{iData$Time}. In the latter case, \code{iData}s will only be generated for the entries in this
vector.}
}
\value{
List of \code{iData} data frames
}
\description{
Splits \code{iData} by the \code{Time} column into multiple \code{iData} data frames.
}
\examples{
#

}
