% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tidy_list}
\alias{tidy_list}
\title{Remove empty components from list}
\usage{
tidy_list(l)
}
\arguments{
\item{l}{A list}
}
\value{
List with empty bits removed
}
\description{
Short cut for removing any empty components of a list
}
\examples{
#

}
