% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treat.R
\name{Treat}
\alias{Treat}
\title{Treat outliers}
\usage{
Treat(x, ...)
}
\arguments{
\item{x}{Object to be treated}

\item{...}{arguments passed to or from other methods.}
}
\value{
Treated object plus details.
}
\description{
Generic function for treating outliers using a two-step process. See individual method documentation:
}
\details{
\itemize{
\item \code{\link[=Treat.numeric]{Treat.numeric()}}
\item \code{\link[=Treat.data.frame]{Treat.data.frame()}}
\item \code{\link[=Treat.coin]{Treat.coin()}}
\item \code{\link[=Treat.unbalanced_coin]{Treat.unbalanced_coin()}}
\item \code{\link[=Treat.purse]{Treat.purse()}}
}

See also \code{vignette("treat")}.

This function replaces the now-defunct \code{treat()} from COINr < v1.0.
}
