% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treat.R
\name{treat2.coin}
\alias{treat2.coin}
\title{Treat a data set in a coin for outliers}
\usage{
\method{treat2}{coin}(x, dset, default_specs = NULL, indiv_specs = NULL, out2 = "coin")
}
\arguments{
\item{x}{A coin}

\item{dset}{A named data set available in \code{.$Data}}

\item{default_specs}{Default specifications}

\item{indiv_specs}{Individual specifications}
}
\value{
An updated coin with a new data set \code{.Data$Treated} added, plus analysis information in
\code{.$Analysis$Treated}.
}
\description{
Treat a data set in a coin for outliers
}
\examples{
#
}
