% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treat.R
\name{treat2.purse}
\alias{treat2.purse}
\title{Treat a purse of coins for outliers}
\usage{
\method{treat2}{purse}(x, dset = NULL, default_specs = NULL, indiv_specs = NULL)
}
\arguments{
\item{x}{A purse object}

\item{dset}{The data set to treat in each coin}

\item{default_specs}{Default specifications}

\item{indiv_specs}{Individual specifications}
}
\value{
An updated purse with new treated data sets added at \code{.$Data$Treated} in each coin, plus
analysis information at \code{.$Analysis$Treated}
}
\description{
Treat a purse of coins for outliers
}
\examples{
#
}
