% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ucodes_to_unames}
\alias{ucodes_to_unames}
\title{Convert uCodes to uNames}
\usage{
ucodes_to_unames(coin, uCodes)
}
\arguments{
\item{coin}{A coin}

\item{uCodes}{A vector of uCodes}
}
\value{
Vector of uNames
}
\description{
Convert uCodes to uNames
}
