% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coin_rew8r.R
\name{weights2corr}
\alias{weights2corr}
\title{Recalculate correlations and ranks based on new weights}
\usage{
weights2corr(COINobj, w, agnames)
}
\arguments{
\item{COINobj}{COINobj object}

\item{w}{a vector of indicator weights}

\item{agnames}{Names of aggregation groups to be included in results table}
}
\value{
A list with .$cr is a vector of correlations between each indicator and the index, and
.$dat is a data frame of rankings, with unit code, and index, input and output scores
}
\description{
This is a short cut function which takes a new set of indicator weights, and recalculates the aggregated results
based on these weights. It returns a summary of rankings and the correlations between indicators and index.
}
