% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{widen}
\alias{widen}
\title{Make long df wide}
\usage{
widen(X)
}
\arguments{
\item{X}{a long format data frame}
}
\value{
A wide format data frame
}
\description{
This is a quick function for making a long-format data frame wide. It is limited in scope, assumes
that the input is a data frame with three columns: one of which is numeric, and the other two are
character vectors. The numeric column will be widened, and the other two columns will be used
for row and column names.
}
