% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{WorldDenoms}
\alias{WorldDenoms}
\title{World denomination data}
\format{
A data frame with 249 rows and 7 variables.
}
\source{
\url{https://data.worldbank.org/}
}
\usage{
WorldDenoms
}
\description{
A small selection of common denominator indicators, which includes GDP, Population, Area, GDP per capita
and income group. All data sourced from the World Bank as of Feb 2021 (data is typically from 2019). Note that this is
intended as example data, and it would be a good idea to use updated data from the World Bank when needed. In this
data set, country names have been altered slightly so as to include no accents - this is simply to make it more
portable between distributions.
}
\keyword{datasets}
