% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{write2coin}
\alias{write2coin}
\title{Direct function outputs}
\usage{
write2coin(coin, l, out2, lev1, lev2, lev3 = NULL)
}
\arguments{
\item{coin}{A coin class object}

\item{l}{The list to direct}

\item{out2}{Whether list or attach to coin}

\item{lev1}{Address at first lev of coin}

\item{lev2}{Address at second lev of coin}

\item{lev3}{Address at third lev of coin}
}
\value{
Either a list or an updated coin
}
\description{
Shortcut to be used at end of functions, either attach to coin or output as list.
}
\examples{
#

}
