% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{write_dset}
\alias{write_dset}
\title{Write a named data set to coin}
\usage{
write_dset(coin, x, dset, quietly = FALSE, ignore_class = FALSE)
}
\arguments{
\item{coin}{A coin class object}

\item{x}{The data to write}

\item{dset}{A character string for naming the data, e.g. \code{Raw}.}

\item{quietly}{If \code{TRUE}, suppresses messages.}

\item{ignore_class}{If \code{TRUE} ignores the class of the input (used for \code{\link[=new_coin]{new_coin()}}).}
}
\value{
Updated coin
}
\description{
Writes a data set to the coin, and performs some checks in the process.
}
\examples{
#

}
