% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{write_log}
\alias{write_log}
\title{Write function arguments to log}
\usage{
write_log(coin, dont_write = NULL)
}
\arguments{
\item{coin}{A coin class object}
}
\value{
Updated GII2 object with function arguments written to \code{.$Log}
}
\description{
This used inside \verb{build_*} functions. It takes the coin object as an input, then writes the arguments of the
current function into the \code{.$Log} list of the coin object. This is then used as a record of the operations used
to build the coin, and can be edited.
}
\examples{
#

}
