% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalise.R
\name{zscore}
\alias{zscore}
\title{Z-score a vector}
\usage{
zscore(x, m_sd = c(0, 1))
}
\arguments{
\item{x}{A numeric vector}

\item{m_sd}{A vector \code{c(m, sd)}, where \code{m} is desired mean and \code{sd} is the target standard deviation.}
}
\value{
Numeric vector
}
\description{
Standardises a vector \code{x} by scaling it to have a mean and standard deviation specified by \code{m_sd}.
}
\examples{
x <- runif(20)
zscore(x)

}
