% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coin_aggregate.R
\name{aggregate}
\alias{aggregate}
\title{Aggregate indicators}
\usage{
aggregate(
  COIN,
  agtype = "arith_mean",
  agweights = NULL,
  dset = "Normalised",
  agtype_bylevel = NULL,
  out2 = NULL
)
}
\arguments{
\item{COIN}{COIN object}

\item{agtype}{The type of aggregation method.}

\item{agweights}{The weights to use in the aggregation. This can either be:
NULL, in which case it will use the weights that were attached to IndMeta and AggMeta in GII_assemble (if they exist), or
A character string which corresponds to a named list of weights stored in .$Parameters$Weights. You can either add these manually or through rew8r.
E.g. entering agweights = "Original" will use the original weights read in on assembly. This is equivalent to agweights = NULL.
Or, a list of weights to use in the aggregation.}

\item{dset}{Which data set (contained in COIN object) to use}

\item{agtype_bylevel}{A character vector with aggregation types for each level}
}
\value{
An updated COIN object containing the new aggregated data set.
}
\description{
Takes indicator data and a specified structure and hierarchically aggregates to a single index (or whatever the structure specified)
}
\examples{
\dontrun{COINlist <- aggregate(COINlist, agtype="arith_mean", dset = "normalised")}

}
