% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{Aggregate}
\alias{Aggregate}
\title{Aggregate data}
\usage{
Aggregate(x, ...)
}
\arguments{
\item{x}{Object to be aggregated}

\item{...}{Further arguments to be passed to methods.}
}
\value{
An object similar to the input
}
\description{
Methods for aggregating numeric vectors, data frames, coins and purses.
}
\examples{
#

}
