% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coin_assemble.R
\name{assemble}
\alias{assemble}
\title{Build COIN object}
\usage{
assemble(
  IndData,
  IndMeta,
  AggMeta,
  include = NULL,
  exclude = NULL,
  preagg = NULL
)
}
\arguments{
\item{IndData}{A dataframe of indicator data.}

\item{IndMeta}{A dataframe containing auxilliary information for each indicator}

\item{AggMeta}{A dataframe specifying the names and weights of each aggregation group}

\item{include}{Optional argument specifying a subset of indicator codes to include (default all indicators included)}

\item{exclude}{Optional argument specifying a subset of indicator codes to exclude (default none excluded)}

\item{preagg}{Set to TRUE if you want to assemble a COIN using pre-aggregated data (typically for ex-post analysis)}
}
\value{
A "COIN object" (list) formatted to the specifications of COINr.
Note that the COIN object is just a tag. It doesn't impose restrictions on the structure of the list.
}
\description{
This takes the raw data provided by the user and puts it into an list format (COIN object) that is recognised by COINr.
It also checks whether there are any syntax errors in the data provided. Optionally, you can exclude
or include indicators using the "include" and "exclude" arguments. Note that if an indicator is specified in
BOTH include and exclude, it will be excluded.
}
\examples{
\dontrun{COINobj <- assemble(IndData, IndMeta, AggMeta)}

}
