% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_ASEM.R
\name{build_ASEM}
\alias{build_ASEM}
\title{Builds ASEM example}
\usage{
build_ASEM()
}
\value{
COIN object of ASEM index.
}
\description{
A short cut function for building the ASEM COIN. Useful for testing.
}
\examples{
\dontrun{ASEM <- build_ASEM()}

}
