% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\name{build_example_coin}
\alias{build_example_coin}
\title{Build ASEM example coin}
\usage{
build_example_coin(up_to = NULL, quietly = FALSE)
}
\arguments{
\item{up_to}{The point up to which to build the index. If \code{NULL}, builds full index. Else specify a \verb{build_*} function
(as a string) - the index will be built up to and including this function. This option is mainly for helping with
function examples. Example: \code{up_to = "build_normalise"}.}

\item{quietly}{If \code{TRUE}, suppresses all messages.}
}
\value{
coin class object
}
\description{
Shortcut function to build the ASEM example coin, using inbuilt example data
}
\examples{
#

}
