% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_and_get.R
\name{check_dset.coin}
\alias{check_dset.coin}
\title{Check for named data set}
\usage{
\method{check_dset}{coin}(x, dset)
}
\arguments{
\item{x}{A coin class object}

\item{dset}{A character string corresponding to a named data set within \code{coin$Data}. E.g. \code{Raw}.}
}
\value{
Error message if \code{dset} not found.
}
\description{
A helper function to check if a named data set is present. If not, will cause an informative error.
}
\examples{
#
}
