% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_coin.R
\name{check_iData}
\alias{check_iData}
\title{Check iData}
\usage{
check_iData(iData, quietly = FALSE)
}
\arguments{
\item{iData}{A data frame of indicator data.}

\item{quietly}{Set \code{TRUE} to suppress message if input is valid.}
}
\value{
Message if everything ok, else error messages.
}
\description{
Checks the format of iData input to \code{\link[=new_coin]{new_coin()}}.
}
\examples{
check_iData(IndData_2020)

}
