% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coin_plotcorr.R
\name{coin_plotcorr}
\alias{coin_plotcorr}
\title{Static heatmaps of correlation using ggplot2}
\usage{
coin_plotcorr(COINobj, inames = NULL, dset = "Raw", plotgroups = TRUE)
}
\arguments{
\item{dset}{The data set to treat}

\item{COIN}{The COIN object}
}
\value{
A treated data set plus information about how the data was treated.
}
\description{
Correlation heatmaps
}
