% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coin_resultsdash.R
\name{coin_resultsdash}
\alias{coin_resultsdash}
\title{Results visualisation dashboard}
\usage{
coin_resultsdash(COINobj, dset = "Aggregated")
}
\arguments{
\item{COINobj}{The COIN object, or a data frame of indicator data.}

\item{dset}{The data set to plot.}
}
\value{
Interactive visualisation
}
\description{
Generates an interactive visualisation of results. Requires Shiny and an active R session.
}
\examples{
\dontrun{coin_indicatordash(COINobj, inames = NULL, dset = "raw")}

}
