% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coinToolIn.R
\name{COINToolIn}
\alias{COINToolIn}
\title{Import data directly from COIN Tool}
\usage{
COINToolIn(fname, makecodes = FALSE, oldtool = FALSE)
}
\arguments{
\item{fname}{The file name and path to read, e.g. \code{"C:/Documents/COINToolFile.xlsx"}.}

\item{makecodes}{Logical: if \code{TRUE}, will generate short indicator codes based on indicator names,
otherwise if \code{FALSE}, will use COIN Tool indicator codes \code{"Ind.01"}, etc. Currently only does this
for indicators, not aggregation groups.}

\item{oldtool}{Logical: if \code{TRUE}, compatible with old COIN Tool (pre-release, early 2019 or earlier).
There are some minor differences on where the elements are found.}
}
\value{
A list
}
\description{
This provides a direct interface for reading a COIN Tool input deck and converting it to COINr.
You need to provide a COIN Tool file, with the "Database" sheet properly compiled.
}
\details{
The \href{https://knowledge4policy.ec.europa.eu/composite-indicators/coin-tool_en}{COIN Tool} is an Excel-based tool for building composite indicators.

See \href{https://bluefoxr.github.io/COINrDoc/helper-functions.html#r-interfaces}{COINr online documentation} for more details and an example.
}
\seealso{
\itemize{
\item \code{\link[=coin2Excel]{coin2Excel()}} Export COIN contents to Excel
}
}
