% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/denominate.R
\name{Denominate.coin}
\alias{Denominate.coin}
\title{Denominate data set in a coin}
\usage{
\method{Denominate}{coin}(
  x,
  dset,
  denoms = NULL,
  denomby = NULL,
  denoms_ID = NULL,
  f_denom = NULL,
  write_to = NULL,
  out2 = "coin"
)
}
\arguments{
\item{x}{A coin class object}

\item{dset}{A named data set to be denominated that is present in .$Data}

\item{denoms}{A data frame of denominator data. If not specified, will extract any potential denominator columns
that were attached to iData when calling new_coin().}

\item{denomby}{Optional data frame specifying which indicators should be denominated and by what. If not specified,
looks for a Denominator column in iMeta.}

\item{denoms_ID}{An ID column for matching denoms with the data to be denominated. This column should contain
uMeta codes to match with the data set extracted from the coin.}

\item{f_denom}{A function which takes two numeric vector arguments and is used to perform the denomination for each
column. By default, this is division, i.e. \code{x[[col]]/denoms[[col]]} for given columns, but any function can be passed
that takes two numeric vectors as inputs and returns a single numeric vector. See details.}

\item{write_to}{If specified, writes the aggregated data to \code{.$Data[[write_to]]}. Default \code{write_to = "Denominated"}.}

\item{out2}{Either \code{"coin"} (default) to return updated coin or \code{"df"} to output the aggregated data set.}
}
\value{

}
\description{
Denominate data set in a coin
}
\examples{
#
}
