% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/denominate.R
\name{Denominate.purse}
\alias{Denominate.purse}
\title{Denominate a data set}
\usage{
\method{Denominate}{purse}(
  x,
  dset,
  denoms = NULL,
  denomby = NULL,
  denoms_ID = NULL,
  f_denom = NULL,
  write_to = NULL,
  ...
)
}
\arguments{
\item{x}{A purse class object}

\item{dset}{The name of the data set to apply the function to, which should be accessible in \code{.$Data}.}

\item{denoms}{A data frame of denominator data. If not specified, will extract any potential denominator columns
that were attached to \code{iData} when calling \code{\link[=new_coin]{new_coin()}}.}

\item{denomby}{Optional data frame specifying which indicators should be denominated and by what. If not specified,
looks for a Denominator column in iMeta.}

\item{denoms_ID}{An ID column for matching \code{denoms} with the data to be denominated. This column should contain
uMeta codes to match with the data set extracted from the coin.}

\item{f_denom}{A function which takes two numeric vector arguments and is used to perform the denomination for each
column. By default, this is division, i.e. \code{x[[col]]/denoms[[col]]} for given columns, but any function can be passed
that takes two numeric vectors as inputs and returns a single numeric vector. See details.}

\item{write_to}{If specified, writes the aggregated data to \code{.$Data[[write_to]]}. Default \code{write_to = "Denominated"}.}

\item{...}{arguments passed to or from other methods.}
}
\value{
An updated purse
}
\description{
Denominate a data set
}
\examples{
#
}
