% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coin_denominate.R
\name{denominate}
\alias{denominate}
\title{Denominate indicator data sets}
\usage{
denominate(
  obj,
  dset = "Raw",
  specby = "metadata",
  denomby = NULL,
  denominators = NULL,
  out2 = "COIN"
)
}
\arguments{
\item{obj}{COIN object (data frames are not currently supported in this function)}

\item{specby}{Selects the source of the specifications for denomination.
If "metadata", uses the denominator column in .$metadata.
If "user", takes a character vector of denominator codes (one for each indicator, with NA for indicators that should not be denominated, and in the same order as the indicators).}

\item{denomby}{Character vector specifying which indicators to use as denominators. Only used if
specby = "user". For indicators with no denomination, set elements to NA.}

\item{out2}{Where to output the results. If "COIN" (default for COIN input), appends to updated COIN,
otherwise if "df" outputs to data frame.}
}
\value{
An updated COIN object, with new dataset .$Data$Denominated of denominated indicators.
}
\description{
Indicators can be denominated (divided) by other external indicators.
Typically, the aim here is to convert extensive (size-related) variables into intensive variables
(comparable between units of different sizes).
}
\examples{
\dontrun{
COIN <- denominate(COIN, specby="metadata", denomby = NULL)}

}
