% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/denominate.R
\name{Denominate}
\alias{Denominate}
\title{Denominate data}
\usage{
Denominate(x, ...)
}
\arguments{
\item{x}{Object to be denominated}

\item{...}{arguments passed to or from other methods}
}
\value{
See individual method documentation
}
\description{
"Denominates" or "scales" variables by other variables. Typically this is done by dividing extensive variables such as
GDP by a scaling variable such as population, to give an intensive variable (GDP per capita).
}
\details{
See documentation for individual methods:
\itemize{
\item \code{\link[=Denominate.data.frame]{Denominate.data.frame()}}
\item \code{\link[=Denominate.coin]{Denominate.coin()}}
\item \code{\link[=Denominate.purse]{Denominate.purse()}}.
}

This function replaces the now-defunct \code{denominate()} from COINr < v1.0.
}
\examples{
# See individual method documentation

}
