% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/denominate.R
\name{Denominate}
\alias{Denominate}
\title{Denominate data}
\usage{
Denominate(x, ...)
}
\arguments{
\item{x}{Thing}

\item{...}{thing}
}
\value{
message
}
\description{
"Denominates" or "scales" variables by other variables. Typically this is done by dividing extensive variables such as
GDP by a scaling variable such as population, to give an intensive variable (GDP per capita).
}
