% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{export_to_excel}
\alias{export_to_excel}
\title{Export a coin or purse to Excel}
\usage{
export_to_excel(x, fname, ...)
}
\arguments{
\item{x}{A coin or purse}

\item{fname}{The file name to write to}

\item{...}{Arguments passed to/from methods}
}
\value{
An Excel spreadsheet.
}
\description{
Writes coins and purses to Excel. See individual method
documentation:
}
\details{
\itemize{
\item \code{\link[=export_to_excel.coin]{export_to_excel.coin()}}
\item \code{\link[=export_to_excel.purse]{export_to_excel.purse()}}
}
}
\examples{
# see individual method documentation

}
