% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlations.R
\name{get_cronbach}
\alias{get_cronbach}
\title{Cronbach's alpha}
\usage{
get_cronbach(coin, ..., use = "pairwise.complete.obs")
}
\arguments{
\item{coin}{A coin or a data frame containing only numerical columns of data.}

\item{...}{Arguments passed to \code{\link[=get_data]{get_data()}}.}

\item{use}{Argument to pass to \link[stats:cor]{stats::cor} to calculate the covariance matrix. Default \code{"pairwise.complete.obs"}.}
}
\value{
Cronbach alpha as a numerical value.
}
\description{
Calculates Cronbach's alpha, a measure of statistical reliability. Cronbach's alpha is a simple measure
of "consistency" of a data set, where a high value implies higher reliability/consistency.
}
\details{
This function simply returns Cronbach's alpha. If you want a lot more details on reliability, the 'psych' package has
a much more detailed analysis.
}
\examples{
#

}
