% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screen_units.R
\name{get_data_avail}
\alias{get_data_avail}
\title{Get data availability of units}
\usage{
get_data_avail(x, ...)
}
\arguments{
\item{x}{Either a coin or a data frame}

\item{...}{Arguments passed to other methods}
}
\description{
Generic function for getting the data availability of each unit (row).
}
\details{
See method documentation:
\itemize{
\item \code{\link[=get_data_avail.data.frame]{get_data_avail.data.frame()}}
\item \code{\link[=get_data_avail.coin]{get_data_avail.coin()}}
}

See also vignettes: \code{vignette("analysis")} and \code{vignette("imputation")}.
}
