% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_and_get.R
\name{get_dset.purse}
\alias{get_dset.purse}
\title{Gets a named data set and performs checks}
\usage{
\method{get_dset}{purse}(x, dset)
}
\arguments{
\item{x}{A purse class object}

\item{dset}{A character string corresponding to a named data set within each coin \code{coin$Data}. E.g. \code{Raw}.}
}
\value{
Data frame of indicator data.
}
\description{
A helper function to retrieve a named data set from a purse object. Retrieves the specified data set
from each coin in the purse and joins them together in a single data frame using \code{\link[=rbind]{rbind()}}, indexed
with a \code{Time} column.
}
\examples{
#

}
