% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_and_get.R
\name{get_dset}
\alias{get_dset}
\title{Gets a named data set and performs checks}
\usage{
get_dset(x, dset, ...)
}
\arguments{
\item{x}{A coin or purse}

\item{dset}{A character string corresponding to a named data set within \code{.$Data}. E.g. \code{"Raw"}.}

\item{...}{arguments passed to or from other methods.}
}
\value{
Data frame of indicator data, indexed also by time if input is a purse.
}
\description{
A helper function to retrieve a named data set from coin or purse objects. See individual
documentation on:
}
\details{
\itemize{
\item \code{\link[=get_dset.coin]{get_dset.coin()}}
\item \code{\link[=get_dset.purse]{get_dset.purse()}}
}
}
\examples{
# see examples for methods

}
