% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights.R
\name{get_eff_weights}
\alias{get_eff_weights}
\title{Get effective weights}
\usage{
get_eff_weights(coin, out2 = "df")
}
\arguments{
\item{coin}{A coin class object}

\item{out2}{Either \code{"coin"} or \code{"df"}}
}
\value{
Either an iMeta data frame with effective weights as an added column, or an updated coin with effective
weights added to \code{.$Meta$Ind}.
}
\description{
Get effective weights
}
