% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA.R
\name{get_PCA}
\alias{get_PCA}
\title{Perform PCA on a coin}
\usage{
get_PCA(
  coin,
  dset = "Raw",
  iCodes = NULL,
  Level = NULL,
  by_groups = TRUE,
  nowarnings = FALSE,
  out2 = "coin"
)
}
\arguments{
\item{coin}{A coin}

\item{dset}{The name of the data set in \code{.$Data} to use.}

\item{iCodes}{An optional character vector of indicator codes to subset the indicator data, passed to \code{\link[=get_data]{get_data()}}}

\item{Level}{The aggregation level to take indicator data from. Integer from 1 (indicator level)
to N (top aggregation level, typically the index).}

\item{by_groups}{If \code{TRUE} (default), performs PCA inside each aggregation group inside the specified level. If \code{FALSE},
performs a single PCA over all indicators/aggregates in the specified level.}

\item{nowarnings}{If \code{FALSE} (default), will give warnings where missing data are found. Set to \code{TRUE} to suppress these warnings.}

\item{out2}{If the input is a coin object, this controls where to send the output. If \code{"coin"}, it
sends the results to the coin object, otherwise if \code{"list"}, outputs to a separate list.}
}
\value{
If \code{out2 = "coin"}, results are appended to the coin object. Specifically:
\itemize{
\item A new set of PCA weights is added to \code{.$Meta$Weights}
\item A list is added to \code{.$Analysis} containing PCA weights (loadings) of the first principle component, and the output of \link[stats:prcomp]{stats::prcomp}, for each
aggregation group found in the targeted level.
If \code{out2 = "list"} the same outputs are contained in a list.
}
}
\description{
Performs Principle Component Analysis (PCA) on a specified data set and subset of indicators or aggregation groups. Returns weights
corresponding to the first principal component, i.e the weights that maximise the variance explained
by the linear combination of indicators.
}
\details{
Note that \code{getPCA()} is simply a quick wrapper for \code{\link[stats:prcomp]{stats::prcomp()}} which makes PCA on COINs quicker.
See \href{https://bluefoxr.github.io/COINrDoc/multivariate-analysis.html#pca}{COINr online documentation} for more details and examples.
}
\examples{
#

}
\seealso{
\itemize{
\item \link[stats:prcomp]{stats::prcomp} Principle component analysis
}
}
