% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity.R
\name{get_sensitivity}
\alias{get_sensitivity}
\title{Sensitivity and uncertainty analysis}
\usage{
get_sensitivity(coin, SA_specs, N, SA_type = "UA")
}
\arguments{
\item{coin}{A coin}

\item{SA_specs}{Specifications of the input uncertainties}

\item{N}{The number of regenerations}

\item{SA_type}{The type of analysis to run. \code{"UA"} runs an uncertainty analysis. \code{"SA"} runs a sensitivity
analysis (which anyway includes an uncertainty analysis).}
}
\value{

}
\description{
Sensitivity and uncertainty analysis
}
\examples{
#
}
