% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{get_stats.coin}
\alias{get_stats.coin}
\title{Get table of statistics for selected indicators}
\usage{
\method{get_stats}{coin}(
  x,
  dset,
  ...,
  t_skew = 2,
  t_kurt = 3.5,
  t_avail = 0.65,
  t_zero = 0.5,
  t_unq = 0.5,
  nsignif = 3,
  out2 = "df"
)
}
\arguments{
\item{x}{A coin}

\item{dset}{A data set present in \code{.$Data}}

\item{...}{Other arguments for indicator selection, passed to \code{\link[=get_data]{get_data()}}.}

\item{t_skew}{Absolute skewness threshold. See details.}

\item{t_kurt}{Kurtosis threshold. See details.}

\item{t_avail}{Data availability threshold. See details.}

\item{nsignif}{Number of significant figures to round the output table to.}

\item{out2}{Either \code{"df"} (default) to output a data frame of indicator statistics, or "\code{coin}" to output an
updated coin with the data frame attached under \code{.$Analysis}.}
}
\value{
Either a data frame or updated coin - see \code{out2}.
}
\description{
Get table of statistics for selected indicators
}
\examples{
#

}
