% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{get_stats}
\alias{get_stats}
\title{Statistics of columns/indicators}
\usage{
get_stats(x, ...)
}
\arguments{
\item{x}{Object (data frame or coin)}

\item{...}{Further arguments to be passed to methods.}
}
\value{
A data frame of statistics for each column
}
\description{
Generic function for reports various statistics from a data frame or coin. See method documentation:
}
\details{
\itemize{
\item \code{\link[=get_stats.data.frame]{get_stats.data.frame()}}
\item \code{\link[=get_stats.coin]{get_stats.coin()}}
}

See also \code{vignette("analysis")}.

This function replaces the now-defunct \code{getStats()} from COINr < v1.0.
}
\examples{
# see individual method documentation

}
