% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIn.R
\name{getIn}
\alias{getIn}
\title{Get subsets of indicator data}
\usage{
getIn(obj, dset = "Raw", inames = NULL, aglev = NULL)
}
\arguments{
\item{obj}{An input object. The function can handle either the COIN object, or a data frame.
The data frame should have each column as an indicator, and an optional column "UnitCode" which
specifies the code (or name) of each unit. Any other type of object will return an error.}

\item{dset}{If input is a COIN object, this specifies which data set in .$Data to use.}

\item{inames}{An optional character vector of indicator codes to subset the indicator data. Usefully, can also refer to
an aggregation group name, and data will be subsetted accordingly. NOTE does not work with multiple aggregate group names.}

\item{aglev}{The aggregation level to take indicator data from. Integer from 1 (indicator level)
to N (top aggregation level, typically the index).}
}
\value{
A list with the following entries:
* .$IndCodes  The indicator codes
* .$IndNames  The indicator names (if a COIN object is input)
* .$ind_data A data frame of indicator data, according to the input specifications, including any unit codes, names and groups
* .$ind_data_only A data frame, as above, but without unit codes, names, groups
* .$UnitCodes Unit codes of selected data set
* .$otype Object type (a string: either "COINobj" or "df")
}
\description{
This function does a number of things that are useful for many COINr functions and operations.
First, it checks to see what kind of input object is input. Then, it selects the indicator data
according to the specs supplied.
}
\details{
For example, specifying dset = "Raw" and inames = c("Ind1", "Ind5"), it will the indicator columns
named "Ind1" and "Ind5" (if they exist), in the format described below. inames can be indicators
or aggregation groups, and can call multiple groups.

You can also specify which aggregation level to target, using the "aglev" argument. See examples
below.
}
\examples{
\dontrun{ # Get data from indicators "Ind1" and "Ind5", from the "Raw" data set
out <- getIn(COINobj, dset = "Raw", inames = c("Ind1", "Ind5"))}
\dontrun{ # get data from "Research" and "Education" dimensions, calling agggregation level 2.
out <- coin_aux_objcheck(COINobj, dset = "Aggregated", inames = c("Research", "Education"), aglev = 2)}

}
