% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coin_preanalyse.R
\name{getStats}
\alias{getStats}
\title{Get table of indicator statistics for any data set}
\usage{
getStats(
  COINobj,
  icodes = NULL,
  dset = "Raw",
  out2 = "COIN",
  t_skew = 2,
  t_kurt = 3.5,
  t_colin = 0.9,
  t_denom = 0.7,
  t_missing = 65,
  IQR_coef = 1.5
)
}
\arguments{
\item{COINobj}{A list of indicator data, stuctured using the COIN_assemble function}

\item{icodes}{A character vector of indicator names to analyse. Defaults to all indicators.}

\item{dset}{The data set to analyse}

\item{out2}{Where to output the results: if "COIN" (default), outputs to the COIN, otherwise if "list", outputs to a separate list.}

\item{t_skew}{Skewness threshold}

\item{t_kurt}{Kurtosis threshold}

\item{t_colin}{Collinearity threshold (absolute value of correlation)}

\item{t_denom}{High correlation with denominator threshold}

\item{t_missing}{Missing data threshold, in percent}

\item{IQR_coef}{Interquartile range coefficient, used for identifying outliers}
}
\value{
If the input is a COIN object, returns an updated COIN object with relevant tables. If the input is a data frame, returns a table as a data frame.
}
\description{
Takes a COIN object, or data frame and returns a table of statistics, including max, min, median, mean, std, kurtosis, etc. Flags indicators with possible outliers.
}
\examples{
\dontrun{df_norm <- getStats(COINobj, dset = "Raw")}

}
