% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coin_aggregate.R
\name{harMean}
\alias{harMean}
\title{Weighted harmonic mean}
\usage{
harMean(x, w = NULL)
}
\arguments{
\item{x}{A numeric vector of positive values.}

\item{w}{A vector of weights, which should have length equal to length(x). Weights are relative
and will be rescaled to sum to 1. If w is not specified, defaults to equal weights.}
}
\value{
Harmonic mean
}
\description{
Weighted harmonic mean of a vector. NAs are skipped by default. This function is used inside
\code{aggregate()}.
}
\examples{
\dontrun{
x <- 1:10
w <- runif(10)
harMean(x,w)}

}
