% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{Impute.numeric}
\alias{Impute.numeric}
\title{Impute a numeric vector}
\usage{
\method{Impute}{numeric}(x, f_i = NULL, f_i_para = NULL, ...)
}
\arguments{
\item{x}{A numeric vector, possibly with \code{NA} values to be imputed.}

\item{f_i}{A function that imputes missing values in a numeric vector}

\item{f_i_para}{Optional further arguments to be passed to \code{f_i()}}

\item{...}{arguments passed to or from other methods.}
}
\value{
An imputed numeric vector of the same length of \code{x}.
}
\description{
This calls the function \code{f_i()}, with optionally further arguments \code{f_i_para}, to impute any missing
values found in \code{x}. By default, \code{f_i = "i_mean()"}, which simply imputes \code{NA}s with the mean of the
non-\code{NA} values in \code{x}.
}
\details{
You could also use one of the imputation functions directly (such as \code{\link[=i_mean]{i_mean()}}). However, this
function offers a few extra advantages, such as checking the input and output formats, and making
sure the resulting imputed vector agrees with the input. It will also skip imputation entirely if
there are no \code{NA}s at all.
}
