% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coin_impute.R
\name{impute}
\alias{impute}
\title{Impute missing data}
\usage{
impute(
  COIN,
  imtype = "ind_mean",
  dset = "Raw",
  groupvar = NULL,
  byyear = FALSE,
  EMaglev = NULL,
  out2 = "COIN"
)
}
\arguments{
\item{COIN}{A COIN or a data frame}

\item{imtype}{The type of imputation method. Either "agg_mean" (the mean of normalised indicators inside the aggregation group),
"agg_median" (the mean of normalised indicators inside the aggregation group),
"ind_mean" (the mean of all the other units in the indicator),
"ind_median" (the median of all the other units in the indicator),
"indgroup_mean" (the mean of all the other units in the indicator, in the same group),
"indgroup_median" (the median of all the other units in the indicator, in the same group),
"EM" (expectation maximisation algorithm via AMELIA package, currently without bootstrapping)}

\item{dset}{The data set in .$Data to impute}

\item{groupvar}{The name of the column to use for by-group imputation. Only applies when imtype is set to a group option.}

\item{byyear}{Logical: set to TRUE to impute separately for each year, otherwise FALSE to impute across all years. NOTE this option is experimental and not tested. Recommend not to use (yet).}

\item{EMaglev}{The aggregation level to use if imtype = "EM".}

\item{out2}{Where to output the results. If "COIN" (default for COIN input), appends to updated COIN,
otherwise if "df" outputs to data frame.}
}
\value{
A dataframe of normalised indicators.
}
\description{
Imputation of data sets using a variety of methods.
}
\examples{
\dontrun{COIN <- coin_impute(COIN, imtype = "ind_mean", dset = "raw")}

}
