% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{Impute}
\alias{Impute}
\title{Impute indicator data sets}
\usage{
Impute(x, ...)
}
\arguments{
\item{x}{Object to be imputed}

\item{...}{arguments passed to or from other methods.}
}
\value{
An object of the same class as \code{x}, but imputed.
}
\description{
Impute indicator data sets
}
\examples{
#

}
