% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{Impute}
\alias{Impute}
\title{Imputation of missing data}
\usage{
Impute(x, ...)
}
\arguments{
\item{x}{Object to be imputed}

\item{...}{arguments passed to or from other methods.}
}
\value{
An object of the same class as \code{x}, but imputed.
}
\description{
This is a generic function with the following methods:
}
\details{
\itemize{
\item \code{\link[=Impute.numeric]{Impute.numeric()}}
\item \code{\link[=Impute.data.frame]{Impute.data.frame()}}
\item \code{\link[=Impute.coin]{Impute.coin()}}
\item \code{\link[=Impute.purse]{Impute.purse()}}
}

See those methods for individual documentation.

This function replaces the now-defunct \code{impute()} from COINr < v1.0.
}
\examples{
# See individual method documentation

}
