% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_coin.R
\name{impute_panel}
\alias{impute_panel}
\title{Impute panel data}
\usage{
impute_panel(iData, time_col = NULL, unit_col = NULL, max_time = NULL)
}
\arguments{
\item{iData}{A data frame of indicator data, containing a time index column \code{time_col}, a unit code column \code{unit_col},
and other numerical columns to be imputed.}

\item{time_col}{The name of a column found in \code{iData} to be used as the time index column. Must point to a numeric column.}

\item{unit_col}{The name of a column found in \code{iData} to be used as the unit code/ID column. Must point to a character column.}

\item{max_time}{The maximum number of time points to look backwards to impute from. E.g. if \code{max_time = 1}, if an
\code{NA} is found at time $t$, it will only look for a replacement value at $t-1$ but not in any time points before that.
By default, searches all time points available.}
}
\value{
A list containing:
\itemize{
\item \code{.$iData_imp}: An \code{iData} format data frame with missing data imputed using previous time points (where possible).
\item \code{.$DataT}: A data frame in the same format as \code{iData}, where each entry shows which time point each data point
came from.
}
}
\description{
Given a data frame of the \code{iData} format, with a time-index column and a unit ID column \code{unit_col}, imputes the
remaining numeric columns using the latest available time point. This function is optionally called inside \code{\link[=new_coin]{new_coin()}}.
}
\details{
This presumes that there are multiple observations for each unit code, i.e. one per time point. It then searches for any missing values in the target year, and replaces them with the equivalent points
from previous time points. It will replace using the most recently available point.
}
\examples{
#

}
