% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coin_resultsdash.R
\name{iplot_bar}
\alias{iplot_bar}
\title{Bar chart}
\usage{
iplot_bar(COINobj, dset = "Raw", isel, usel)
}
\arguments{
\item{COINobj}{The COIN object, or a data frame of indicator data.}

\item{dset}{The data set to plot.}

\item{isel}{The selected indicator code or aggregate}
}
\value{
Interactive bar chart.
}
\description{
Generates an interactive bar chart. This function is simply a wrapper for
the Plotly bar chart function, but accesses COIN object to get the relevant indicator.
Also has click event data for Shiny.
}
\examples{
\dontrun{iplot_bar(COINobj, dset = "Aggregated", inames = "Index")}

}
