% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coin_resultsdash.R
\name{iplotBar}
\alias{iplotBar}
\title{Bar chart}
\usage{
iplotBar(COINobj, dset = "Raw", isel = NULL, usel = NULL, aglev = 1)
}
\arguments{
\item{COINobj}{The COIN object, or a data frame of indicator data.}

\item{dset}{The data set to plot.}

\item{isel}{The selected indicator code or aggregate (does not support multiple indicators)}

\item{usel}{A character vector of unit codes to highlight on the bar chart}

\item{aglev}{The aggregation level to collect the indicator data from}
}
\value{
Interactive bar chart.
}
\description{
Generates an interactive bar chart. This function is simply a wrapper for
the Plotly bar chart function, but accesses COIN object to get the relevant indicator.
Also has click event data for Shiny.
}
\examples{
\dontrun{iplotBar(ASEM, dset = "Raw", isel = "Flights")}

}
