% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coin_resultsdash.R
\name{iplotBar}
\alias{iplotBar}
\title{Bar chart}
\usage{
iplotBar(
  COIN,
  dset = "Raw",
  isel = NULL,
  usel = NULL,
  aglev = NULL,
  stack_children = FALSE
)
}
\arguments{
\item{COIN}{The COIN object, or a data frame of indicator data.}

\item{dset}{The data set to plot.}

\item{isel}{The selected indicator code or aggregate (does not support multiple indicators)}

\item{usel}{A character vector of unit codes to highlight on the bar chart}

\item{aglev}{The aggregation level to collect the indicator data from}

\item{stack_children}{If TRUE, produces a stacked bar chart with any children of isel. In this case, usel is ignored.
This only works if dset = "Aggregated" and aglev > 1.}
}
\value{
Interactive bar chart.
}
\description{
Generates an interactive bar chart. This function is simply a wrapper for
the Plotly bar chart function, but accesses COIN object to get the relevant indicator.
Also has click event data for Shiny.
}
\examples{
\dontrun{iplotBar(ASEM, dset = "Raw", isel = "Flights")}

}
