% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.coin}
\alias{is.coin}
\title{Check if object is coin class}
\usage{
is.coin(x)
}
\arguments{
\item{x}{An object to be checked.}
}
\value{
Logical
}
\description{
Check if object is coin class
}
\examples{
is.coin("not_a_coin")
# add example with coin

}
