% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIn.R
\name{is.COIN}
\alias{is.COIN}
\title{Check if an object is a COIN}
\usage{
is.COIN(obj)
}
\arguments{
\item{obj}{An input object to test}
}
\value{
Logical: TRUE if input is a COIN, otherwise FALSE
}
\description{
Returns TRUE if an input object is a COIN, otherwise FALSE if not.
}
\examples{
# build the ASEM COIN
ASEM <- assemble(IndData = ASEMIndData, IndMeta = ASEMIndMeta, AggMeta = ASEMAggMeta)
# check class
stopifnot(is.COIN(ASEM))

}
\seealso{
\itemize{
\item \code{\link[=getIn]{getIn()}} Get subset of indicator data from either a COIN or data frame.
\item \code{\link[=assemble]{assemble()}} Assemble a COIN from indicator data and metadata
}
}
