% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coin_normalise.R
\name{normalise}
\alias{normalise}
\title{Normalise indicator data sets}
\usage{
normalise(
  COINobj,
  ntype = "minmax",
  npara = NULL,
  inames = NULL,
  dset = "Raw",
  directions = NULL,
  out2 = NULL
)
}
\arguments{
\item{COINobj}{Either the COIN object, or a data frame of indicator data}

\item{ntype}{The type of normalisation method. Either "minmax", "zscore", or "custom".}

\item{npara}{Supporting object for ntype.}

\item{inames}{Character vector of indicator names, indicating which columns to normalise. Use this if you only want to normalise certain columns, or you are inputting a data frame with some columns which are not to be normalised (e.g. country names, groups, ...)}

\item{dset}{The data set to normalise}

\item{directions}{A vector specifying the direction assigned to each indicator.
Needs to be the same length as the number of indicators, or the number of indicators in inames, if specified.}

\item{out2}{Where to output the results. If "COIN" (default for COIN input), appends to updated COIN,
otherwise if "df" outputs to data frame.}
}
\value{
An updated COIN object with .$Data$Normalised added.
}
\description{
A dataset of indicators is normalised using one of several methods. This function also supports custom normalisation.
}
\examples{
\dontrun{df_norm <- normalise(COINobj, ntype="minmax", npara = c(0,1))}

}
