% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coin_normalise.R
\name{normalise}
\alias{normalise}
\title{Normalise indicator data sets}
\usage{
normalise(
  COIN,
  ntype = "minmax",
  npara = NULL,
  dset = "Raw",
  directions = NULL,
  individual = NULL,
  indiv_only = FALSE,
  out2 = NULL
)
}
\arguments{
\item{COIN}{Either the COIN object, or a data frame of indicator data}

\item{ntype}{The type of normalisation method. Either "minmax", "zscore", "scaled", "rank", "borda", "prank", "fracmax", "dist2targ", "dist2ref", "dist2max", "custom" or "none".}

\item{npara}{Supporting object for ntype.}

\item{dset}{The data set to normalise}

\item{directions}{A vector specifying the direction assigned to each indicator.
Needs to be the same length as the number of indicators, or the number of indicators in icodes, if specified.}

\item{individual}{A list of named lists specifiying individual normalisation to apply to specific indicators. Should be structured as follows:
The name of each sublist should be the indicator code. The the list elements are:
.$ntype is the type of normalisation to apply
.$npara is a corresponding object or parameters that are used by ntype}

\item{indiv_only}{Logical: if FALSE (default), indicators not specified in individual are subjected to default normalisation. Otherwise if TRUE they are not normalised.}

\item{out2}{Where to output the results. If "COIN" (default for COIN input), appends to updated COIN,
otherwise if "df" outputs to data frame.}
}
\value{
An updated COIN object with .$Data$Normalised added.
}
\description{
A dataset of indicators is normalised using one of several methods. This function also supports custom normalisation.
}
\examples{
\dontrun{df_norm <- normalise(COIN, ntype="minmax", npara = c(0,1))}

}
