% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalise.R
\name{Normalise}
\alias{Normalise}
\title{Normalise data}
\usage{
Normalise(x, ...)
}
\arguments{
\item{x}{Object to be normalised}

\item{...}{Further arguments to be passed to methods.}
}
\description{
Normalise data
}
\examples{
#

}
