% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalise.R
\name{Normalise}
\alias{Normalise}
\title{Normalise data}
\usage{
Normalise(x, ...)
}
\arguments{
\item{x}{Object to be normalised}

\item{...}{Further arguments to be passed to methods.}
}
\description{
This is a generic function for normalising variables and indicators, i.e. bringing them onto
a common scale. Please see individual method documentation depending on your data class:
}
\details{
\itemize{
\item \code{\link[=Normalise.numeric]{Normalise.numeric()}}
\item \code{\link[=Normalise.data.frame]{Normalise.data.frame()}}
\item \code{\link[=Normalise.coin]{Normalise.coin()}}
\item \code{\link[=Normalise.purse]{Normalise.purse()}}
}

See also \code{vignette("normalise")} for more details.

This function replaces the now-defunct \code{normalise()} from COINr < v1.0.
}
\examples{
# See individual method documentation.

}
