% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coin_normalise.R
\name{normalise}
\alias{normalise}
\title{Normalise indicator data sets}
\usage{
normalise(
  COIN,
  ntype = "minmax",
  npara = NULL,
  dset = NULL,
  directions = NULL,
  individual = NULL,
  indiv_only = NULL,
  out2 = NULL
)
}
\arguments{
\item{COIN}{Either the COIN object, or a data frame of indicator data}

\item{ntype}{The type of normalisation method. Either "minmax", "zscore", "scaled", "rank", "borda", "prank", "fracmax", "dist2targ", "dist2ref", "dist2max", "custom" or "none".}

\item{npara}{Supporting object for ntype. This is a list of the form list(ntype = parameters_for_ntype). So,
if ntype = "minmax", npara could be list(minmax = c(0, 100)) to scale into the 0 to 100 interval
If ntype = "zscore", npara could be list(zscore = c(0, 1)) to scale to mean zero and std 1.
This means you can store parameters for more than one normalisation type side by side, which helps in
comparisons, adjustments, and sensitivity analyses.}

\item{dset}{The data set to normalise}

\item{directions}{A vector specifying the direction assigned to each indicator.
Needs to be the same length as the number of indicators, or the number of indicators in icodes, if specified.}

\item{individual}{A list of named lists specifiying individual normalisation to apply to specific indicators. Should be structured as follows:
The name of each sublist should be the indicator code. The the list elements are:
.$ntype is the type of normalisation to apply
.$npara is a corresponding object or parameters that are used by ntype, in the same format as npara above.}

\item{indiv_only}{Logical: if FALSE (default), indicators not specified in individual are subjected to default normalisation. Otherwise if TRUE they are not normalised.}

\item{out2}{Where to output the results. If "COIN" (default for COIN input), appends to updated COIN,
otherwise if "df" outputs to data frame.}
}
\value{
An updated COIN object with .$Data$Normalised added, or a normalised data frame.
}
\description{
A dataset of indicators is normalised using one of several methods. This function also supports custom normalisation.
}
\details{
Normalisation refers to the operation of bringing variables (indicators) onto a common scale. This is typically done by matching
one or more indicator statistics. For example, the \emph{min-max} method operates a linear transformation to make the minimum and
maximum values of each indicator to be equal. The \emph{z-score} method makes the standard deviation and variance equal. And so on.

This function supports a range of normalisation methods - see \code{ntype}. Some of these require supporting parameters or similar -
to see full details check the \href{https://bluefoxr.github.io/COINrDoc/normalisation.html}{online documentation}.

Indicators can also be each normalised by a different method. See \code{individual}.
}
\examples{
\dontrun{
# build ASEM COIN
ASEM <- assemble(IndData = ASEMIndData, IndMeta = ASEMIndMeta, AggMeta = ASEMAggMeta)
# directly normalise raw data using min-max, onto 0-10 interval
ASEM <- normalise(ASEM, dset = "Raw", ntype = "minmax", npara = list(minmax = c(0,10)))}

}
