% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coin_aggregate.R
\name{outrankMatrix}
\alias{outrankMatrix}
\title{Outranking matrix}
\usage{
outrankMatrix(ind_data, w = NULL)
}
\arguments{
\item{ind_data}{A data frame or matrix of indicator data, with observations as rows and indicators
as columns. No other columns should be present (e.g. label columns).}

\item{w}{A vector of weights, which should have length equal to ncol(ind_data). Weights are relative
and will be rescaled to sum to 1. If w is not specified, defaults to equal weights.}
}
\value{
Outranking matrix
}
\description{
Constructs an outranking matrix based on a data frame of indicator data and corresponding weights
}
