% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bar.R
\name{plot_bar}
\alias{plot_bar}
\title{Bar chart}
\usage{
plot_bar(
  coin,
  dset,
  iCode,
  ...,
  uLabel = "uCode",
  axes_label = "iCode",
  by_group = NULL,
  dset_label = FALSE,
  log_scale = FALSE
)
}
\arguments{
\item{coin}{A coin object.}

\item{dset}{Data set from which to extract the variable to plot. Passed to \code{\link[=get_data]{get_data()}}.}

\item{iCode}{Code of variable or indicator to plot. Passed to \code{\link[=get_data]{get_data()}}.}

\item{...}{Further arguments to pass to \code{\link[=get_data]{get_data()}}, e.g. for filtering units.}

\item{uLabel}{How to label units: either \code{"iCode"}, or \code{"iName"}.}

\item{axes_label}{How to label the y axis and group legend: either \code{"uCode"} or \code{"uName"}.}

\item{by_group}{Optional group variable to use to colour bars.}

\item{dset_label}{Logical: whether to include the data set in the y axis label.}

\item{log_scale}{Logical: if \code{TRUE} uses a log scale for the y axis.}
}
\value{
A ggplot2 plot object.
}
\description{
Plot bar charts of single indicators, possibly coloured according to a grouping variable.
}
\examples{
#
}
