% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dist.R
\name{plot_dot}
\alias{plot_dot}
\title{Dot plots of single indicator with highlighting}
\usage{
plot_dot(
  coin,
  ...,
  usel = NULL,
  marker_type = "circle",
  add_stat = NULL,
  stat_label = NULL,
  show_ticks = TRUE,
  plabel = NULL,
  usel_label = TRUE,
  vert_adjust = 0.5
)
}
\arguments{
\item{coin}{The coin}

\item{...}{Arguments to pass to \code{\link[=get_data]{get_data()}}. Note that \code{iCodes} is required to be of length 1.}

\item{usel}{A subset of units to highlight.}

\item{marker_type}{The type of marker, either \code{"circle"} (default) or \code{"cross"}, or a marker number to pass to ggplot2 (0-25).}

\item{add_stat}{A statistic to overlay, either \code{"mean"}, \code{"median"} or else a specified value.}

\item{stat_label}{An optional string to use as label at the point specified by \code{add_stat}.}

\item{show_ticks}{Set \code{FALSE} to remove axis ticks.}

\item{plabel}{Controls the labelling of the indicator. If not specified, returns the indicator name,
plus units if found. Otherwise if \code{"indname"}, returns only indicator name, if \code{"indname+unit"}, returns
indicator name plus unit (if found), if \code{"unit"} returns only unit (if found), otherwise if \code{"none"},
displays no text. Finally, any other string can be passed, so e.g. \code{"My indicator"} will display this on the
axis.}

\item{usel_label}{If \code{TRUE} (default) also labels selected units with their unit codes. \code{FALSE} to disable.}

\item{vert_adjust}{Adjusts the vertical height of text labels and stat lines, which matters depending on plot size.
Takes a value between 0 to 2 (higher will probably remove the label from the axis space).}
}
\value{
A ggplot2 plot object.
}
\description{
Plots a single indicator as a line of dots, and optionally highlights a selected unit.
}
\examples{
#

}
