% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coin_rew8r.R
\name{plotCorr}
\alias{plotCorr}
\title{Correlation heatmap}
\usage{
plotCorr(
  COIN,
  aglevs = NULL,
  insig = FALSE,
  levs = TRUE,
  grouprects = TRUE,
  mapcols = "thresholds",
  corthresh = NULL,
  showvals = TRUE,
  cortype = "pearson",
  useweights = NULL
)
}
\arguments{
\item{COIN}{The COIN object}

\item{aglevs}{A two length vector specifying which level to plot against which level. E.g. c(2,4) for
COIN plots sub-pillars against sub-indexes. If NULL, plots everything against everything.}

\item{insig}{Logical: if TRUE, all correlation values are plotted; if FALSE (default), does not plot insignificant correlations.}

\item{levs}{Logical: if TRUE, plots lines showing the division between different levels. Only works if aglevs = NULL.}

\item{grouprects}{Logical: if TRUE, plots rectangles showing aggregation groups}

\item{mapcols}{If set to "thresholds", uses a discrete colour scale specified by corthresh. Otherwise if "continuous"
uses a continuous colour map.}

\item{corthresh}{A named list specifying the colour thresholds to use if mapcols = "thresholds". Entries should
specify correlation thresholds and can specify any of clow, cmid and chi. Anything below clow will be
coloured red. Anything between clow and cmid will be grey. Anything between cmid and chigh will be blue.
Anything above chigh will be green. Default is list(clow = -0.4, cmid = 0.4, chigh = 0.85). You can
specify a subset of these and the others will revert to defaults.}

\item{showvals}{Logical: if TRUE, overlays correlation values on each square.}

\item{cortype}{The type of correlation: either "pearson" (default), "spearman" or "kendall"}

\item{useweights}{An optional list of weights to use (this is used mainly in the rew8r app)}
}
\value{
A correlation map
}
\description{
Plots a heatmap of a correlation matrix
}
\examples{
\dontrun{flags <- plotCorr(COIN, aglevs = c(2,3))}

}
