% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coin_multivariate.R
\name{plotCorr}
\alias{plotCorr}
\title{Static heatmaps of correlation using ggplot2}
\usage{
plotCorr(COIN, dset = "Raw", icodes = NULL, aglevs = 1)
}
\arguments{
\item{COIN}{The COIN object}

\item{dset}{The data set to treat}

\item{icodes}{An optional list or character vector where the first entry specifies the indicator/agg
codes to correlate against the second entry (also a specification of ind/agg codes)}

\item{aglevs}{The aggregation levels to take the two groups of indicators from. See getIn() for details.}
}
\value{
A treated data set plus information about how the data was treated.
}
\description{
Correlation heatmaps
}
