% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coin_sensitivity.R
\name{plotSA}
\alias{plotSA}
\title{Plot sensitivity indices}
\usage{
plotSA(SAresults, ptype = "bar")
}
\arguments{
\item{SAresults}{A list of sensitivity/uncertainty analysis results from COINr::sensitivity().}

\item{ptype}{Type of plot to generate - either "bar", "pie" or "box"}
}
\description{
Plots sensitivity indices as bar or pie charts.
To use this function you first need to run sensitivity().
}
