% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIn.R
\name{print.COIN}
\alias{print.COIN}
\title{Print COIN}
\usage{
\method{print}{COIN}(x, ...)
}
\arguments{
\item{x}{A COIN}

\item{...}{Arguments to be passed to or from other methods.}
}
\value{
Text output
}
\description{
Some details about the COIN
}
\examples{
ASEM <- build_ASEM()
print(ASEM)

}
