% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qfuncs.R
\name{qNormalise.coin}
\alias{qNormalise.coin}
\title{Quick normalisation of a coin}
\usage{
\method{qNormalise}{coin}(
  x,
  dset,
  f_n = "n_minmax",
  f_n_para = list(l_u = c(0, 100)),
  directions = NULL,
  ...
)
}
\arguments{
\item{x}{A coin}

\item{dset}{Name of data set to normalise}

\item{f_n}{Name of a normalisation function (as a string) to apply to each indicator. Default \code{"n_minmax"}.}

\item{f_n_para}{Any further arguments to pass to \code{f_n}, as a named list.}

\item{directions}{An optional data frame containing the following columns:
\itemize{
\item \code{iCode} The indicator code, corresponding to the column names of the data frame
\item \code{Direction} numeric vector with entries either \code{-1} or \code{1}
If \code{directions} is not specified, the directions will be taken from the \code{iMeta} table in the coin, if available.
}}

\item{...}{arguments passed to or from other methods.}
}
\value{
An updated coin
}
\description{
Quick normalisation of a coin
}
\examples{
#
}
