% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qfuncs.R
\name{qTreat.coin}
\alias{qTreat.coin}
\title{Quick outlier treatment of a coin}
\usage{
\method{qTreat}{coin}(
  x,
  dset,
  winmax = 5,
  skew_thresh = 2,
  kurt_thresh = 3.5,
  f2 = "log_GII",
  ...
)
}
\arguments{
\item{x}{A coin}

\item{dset}{Name of data set to treat for outliers}

\item{winmax}{Maximum number of points to Winsorise for each indicator. Default 5.}

\item{skew_thresh}{Absolute skew threshold - default 2.}

\item{kurt_thresh}{Kurtosis threshold - default 3.5.}

\item{f2}{Function to call if Winsorisation does not bring skew and kurtosis within limits. Default \code{"log_GII"}.}

\item{...}{arguments passed to or from other methods.}
}
\value{
An updated coin
}
\description{
Quick outlier treatment of a coin
}
\examples{
#
}
