% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qfuncs.R
\name{qTreat.data.frame}
\alias{qTreat.data.frame}
\title{Quick outlier treatment of a data frame}
\usage{
\method{qTreat}{data.frame}(x, winmax = 5, skew_thresh = 2, kurt_thresh = 3.5, f2 = "log_GII", ...)
}
\arguments{
\item{x}{A numeric data frame}

\item{winmax}{Maximum number of points to Winsorise for each column. Default 5.}

\item{skew_thresh}{Absolute skew threshold - default 2.}

\item{kurt_thresh}{Kurtosis threshold - default 3.5.}

\item{f2}{Function to call if Winsorisation does not bring skew and kurtosis within limits. Default \code{"log_GII"}.}

\item{...}{arguments passed to or from other methods.}
}
\value{
A list
}
\description{
A simplified version of \code{\link[=Treat]{Treat()}} which allows direct access to the default parameters. This has less flexibility,
but is an easier interface and probably more convenient if the objective is to use the default treatment process
but with some minor adjustments.
}
\examples{
#
}
