% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qfuncs.R
\name{qTreat}
\alias{qTreat}
\title{Quick outlier treatment}
\usage{
qTreat(x, ...)
}
\arguments{
\item{x}{Object to be normalised.}

\item{...}{arguments passed to or from other methods.}
}
\value{
A treated object
}
\description{
This is a generic wrapper function for \code{\link[=Treat]{Treat()}}. It offers a simpler syntax but less flexibility.
}
\details{
See individual method documentation:
\itemize{
\item \code{\link[=qTreat.data.frame]{qTreat.data.frame()}}
\item \code{\link[=qTreat.coin]{qTreat.coin()}}
\item \code{\link[=qTreat.purse]{qTreat.purse()}}
}
}
\examples{
# See individual method examples

}
