% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coin_resultstable.R
\name{rankDF}
\alias{rankDF}
\title{Convert a data frame to ranks}
\usage{
rankDF(df)
}
\arguments{
\item{df}{A data frame}
}
\value{
Data frame with ranks
}
\description{
Replaces all numerical columns of a data frame with their ranks. Uses sport ranking, i.e. ties
share the highest rank place. Ignores non-numerical columns. See \code{\link[=rank]{rank()}}.
}
\examples{
\dontrun{
# some random data
df <- as.data.frame(matrix(runif(20), 5, 4))
# convert to ranks
rankDF(df)}

}
\seealso{
\itemize{
\item \code{\link[=roundDF]{roundDF()}} Round a data frame to a specified number of decimals.
}
}
