% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rbind_fill}
\alias{rbind_fill}
\title{rbind two lists with different names into a data frame}
\usage{
rbind_fill(x1, x2)
}
\arguments{
\item{x1}{A named list or named vector}

\item{x2}{Another named list or named vector}
}
\value{
Data frame
}
\description{
Performs an \code{rbind()} operation on two named lists or vectors that do not need to share the same names, but
will match the names and fill any missing cols with \code{NA}s.
}
\examples{
#

}
