% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regen.R
\name{Regen.coin}
\alias{Regen.coin}
\title{Regenerate a coin}
\usage{
\method{Regen}{coin}(x, from = NULL, quietly = TRUE)
}
\arguments{
\item{x}{A coin class object}

\item{from}{Optional: a construction function name. If specified, regeneration begins from this function, rather
than re-running all functions.}

\item{quietly}{If \code{TRUE} (default), messages are suppressed during building.}
}
\value{
Updated coin object with regenerated results.
}
\description{
Regenerates the \code{.$Data} entries by rerunning the construction functions according to the specifications in \code{.$Log}.
This effectively regenerates the results. Different variations of coins can be quickly achieved by editing the
saved arguments in \code{.$Log} and regenerating.
}
\details{
The \code{from} argument allows partial regeneration, starting from a
specified function. This can be helpful to speed up regeneration in some cases. However, keep in mind that
if you change a \code{.$Log} argument from a function that is run before the point that you choose to start running
from, it will not affect the results.

Note that while sets of weights will be passed to the regenerated COIN, anything in \code{.$Analysis} will be removed
and will have to be recalculated.
}
\examples{
#

}
