% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coin_treat.R
\name{treat}
\alias{treat}
\title{Treatment of outliers}
\usage{
treat(
  COIN,
  dset = "Raw",
  winmax = NULL,
  winchange = FALSE,
  deflog = "log",
  boxlam = NULL,
  t_skew = 2,
  t_kurt = 3.5,
  individual = NULL,
  indiv_only = TRUE
)
}
\arguments{
\item{COIN}{The COIN object}

\item{dset}{The data set to treat}

\item{winmax}{The maximum number of points to Winsorise for each indicator. If NA, will keep Winsorising until skew&kurt thresholds achieved (but it is likely this will cause errors)}

\item{winchange}{Logical: if TRUE, Winsorisation can change direction from one iteration to the next. Otherwise if FALSE (default), no change.}

\item{deflog}{The type of transformation to apply if Winsorisation fails. If "log", use simple ln(x) as log transform (note: indicators containing negative values
will be skipped). If "CTlog", will do ln(x-min(x)+1), as used in the COIN Tool. If "GIIlog", use GII log transformation.
If "boxcox", performs a Box Cox transformation. In this latter case, you should also specify boxlam. Finally, if "none", will
return the indicator untreated.}

\item{boxlam}{The lambda parameter of the Box Cox transform.}

\item{t_skew}{Absolute skew threshold (default 2)}

\item{t_kurt}{Kurtosis threshold (default 3.5)}

\item{individual}{A data frame specifying individual treatment for each indicator, with each row correspoding to one indicator to be treated. Rows are:
"IndCode" The code of the indicator to be treated
"Treat" The type of treatment to apply, one of "win" (Winsorise), "log" (log), "GIIlog" (GII log), "CTlog" (COIN Tool log), "boxcox" (Box Cox), or "None" (no treatment)
"winmax" The maximum number of points to Winsorise. Ignored if the corresponding entry in "Treat" is not "win"
"thresh" Either NA, which means that Winsorisation will continue up to winmax with no checks on skew and kurtosis, or "thresh", which uses the skew and kurtosis thresholds specified in t_skew and t_kurt.
"boxlam" Lambda parameter for the Box Cox transformation}

\item{indiv_only}{Logical: if TRUE, only the indicators specified in "individual" are treated.
If false, all indicators are treated: any outside of "individual" will get default treatment.}
}
\value{
A treated data set plus information about how the data was treated.
}
\description{
Takes the COIN object and Winsorises indicators where necessary or specified, or reverts to log transform.
}
