% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treat.R
\name{treat2.coin}
\alias{treat2.coin}
\title{Treat a data set in a coin for outliers}
\usage{
\method{treat2}{coin}(
  x,
  dset,
  default_specs = NULL,
  indiv_specs = NULL,
  combine_treat = FALSE,
  out2 = "coin"
)
}
\arguments{
\item{x}{A coin}

\item{dset}{A named data set available in \code{.$Data}}

\item{default_specs}{Default specifications}

\item{indiv_specs}{Individual specifications}

\item{combine_treat}{By default, if \code{f1} fails to pass \code{f_pass}, then \code{f2} is applied to the original \code{x},
rather than the treated output of \code{f1}. If \code{combine_treat = TRUE}, \code{f2} will instead be applied to the output
of \code{f1}, so the two treatments will be combined.}
}
\value{
An updated coin with a new data set \code{.Data$Treated} added, plus analysis information in
\code{.$Analysis$Treated}.
}
\description{
Treat a data set in a coin for outliers
}
\examples{
#
}
