% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights.R
\name{weights2corr}
\alias{weights2corr}
\title{Recalculate correlations and ranks based on new weights}
\usage{
weights2corr(
  coin,
  w,
  dset = "Aggregated",
  Levels = NULL,
  cortype = "pearson",
  withparent = TRUE
)
}
\arguments{
\item{coin}{coin object}

\item{w}{Full data frame of weights for each level}

\item{dset}{Name of the data set that is created when \code{\link[=Aggregate]{Aggregate()}} is called. This is used to calculated correlations
and to extract the results table. Default \code{"Aggregated"}.}

\item{Levels}{A 2-length vector with two aggregation levels to correlate against each other}

\item{cortype}{Correlation type. Either \code{"pearson"} (default), \code{"kendall"} or \code{"spearman"}. See \link[stats:cor]{stats::cor}.}

\item{withparent}{Logical: if \code{TRUE}, only correlates with the parent, e.g. sub-pillars are only correlated with
their parent pillars and not others.}
}
\value{
A list where \code{.$cr} is a vector of correlations between each indicator and the index, and
\code{.$dat} is a data frame of results
}
\description{
This is a short cut function which takes a new set of indicator weights, and recalculates the coin results
based on these weights. It returns a summary of rankings and the correlations between indicators and index.
}
\details{
This function is principally used inside \code{\link[=rew8r]{rew8r()}}. The \code{w} argument should be a data frame of weights, of the same format
as the data frames found in \code{.$Parameters$Weights}.
}
\examples{
#
}
